% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientfunctions.R
\name{AdjustFilter}
\alias{AdjustFilter}
\title{Adjust the filter such that it reflects the pooling operation performed.}
\usage{
AdjustFilter(poolingFilter, A.hat, X, Theta)
}
\arguments{
\item{poolingFilter}{A PoolingFilter object.}

\item{A.hat}{Convolutional operator.}

\item{X}{Input matrix from last training iteration.}

\item{Theta}{Weight matrix from last training iteration.}
}
\description{
Adjust the filter such that it reflects the pooling operation performed.
}
