% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientfunctions.R
\name{BackpropagateSingleLayer}
\alias{BackpropagateSingleLayer}
\title{Run backpropagation for a single layer. In backpropagation, a gradient is
is computed by taking partial derivatives for each of the model weights. Given
a learning rate, the weights are adjusted according to the gradient.}
\usage{
BackpropagateSingleLayer(modelResults, iteration)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}

\item{iteration}{The current iteration.}
}
\description{
Run backpropagation for a single layer. In backpropagation, a gradient is
is computed by taking partial derivatives for each of the model weights. Given
a learning rate, the weights are adjusted according to the gradient.
}
