% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_BuildCoRegulationGraph.R
\name{BuildCoRegulationGraph}
\alias{BuildCoRegulationGraph}
\title{Construct a co-regulation graph where nodes are genes and metabolites, and edges
are present between nodes and edges that meet the processing cutoffs. Note that
the ProcessResults function must be run before BuildCoRegulationGraph is run.
This is because the BuildCoRegulationGraph uses the output from ProcessResults
to generate the graph. The graph generated by this function is undirected and
is an igraph object.}
\usage{
BuildCoRegulationGraph(
  inputResults,
  independentVarType,
  outcome,
  vertexSize = 15
)
}
\arguments{
\item{inputResults}{A list of IntLimResults objects. Each object must include
model and processing results (output of ProcessResults()).}

\item{independentVarType}{The independent variable type ("gene" or "metabolite")}

\item{outcome}{The outcome type ("gene" or "metabolite")}

\item{vertexSize}{Width of each vertex in pixels}
}
\description{
Construct a co-regulation graph where nodes are genes and metabolites, and edges
are present between nodes and edges that meet the processing cutoffs. Note that
the ProcessResults function must be run before BuildCoRegulationGraph is run.
This is because the BuildCoRegulationGraph uses the output from ProcessResults
to generate the graph. The graph generated by this function is undirected and
is an igraph object.
}
