% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_BuildCoRegulationGraph.R
\name{BuildCoRegulationGraphAllFolds}
\alias{BuildCoRegulationGraphAllFolds}
\title{Construct a co-regulation graph for each fold.}
\usage{
BuildCoRegulationGraphAllFolds(
  inputResults,
  vertexSize = 15,
  independentVarType,
  outcome
)
}
\arguments{
\item{inputResults}{A list of data frames. Each data frame consists of the
filt.results slot from an IntLIMResults object.}

\item{vertexSize}{Width of each vertex in pixels}

\item{independentVarType}{The independent variable type ("gene" or "metabolite")}

\item{outcome}{The outcome type ("gene" or "metabolite")}
}
\description{
Construct a co-regulation graph for each fold.
}
