% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{BuildDataAndLines}
\alias{BuildDataAndLines}
\title{A helper function for the PlotPair functions (i.e. the highcharter one and
the flat, base-R one).}
\usage{
BuildDataAndLines(
  inputData,
  inputResults,
  outcome,
  independentVariable,
  independentAnalyteOfInterest,
  outcomeAnalyteOfInterest,
  palette = "Set1",
  stype
)
}
\arguments{
\item{inputData}{IntLimObject output of ReadData() or FilterData()}

\item{inputResults}{Data frame with model results (output of ProcessResults())}

\item{outcome}{'1' or '2' must be set as outcome/independent variable}

\item{independentVariable}{'1' or '2' must be set as outcome/independent variable}

\item{independentAnalyteOfInterest}{independent analyte in pair}

\item{outcomeAnalyteOfInterest}{outcome analyte in pair}

\item{palette}{choose an RColorBrewer palette ("Set1", "Set2", "Set3",
"Pastel1", "Pastel2", "Paired", etc.) or submit a vector of colors}

\item{stype}{Phenotype or outcome variable}
}
\description{
A helper function for the PlotPair functions (i.e. the highcharter one and
the flat, base-R one).
}
