% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_BuildCoRegulationGraph.R
\name{BuildGraphDataFrame}
\alias{BuildGraphDataFrame}
\title{Construct a data frame that includes all information needed to build an igraph
object. This includes the names of the two co-regulated analytes, whether the
analytes have a positive or negative co-regulation, and the shape of the analytes.}
\usage{
BuildGraphDataFrame(inputResults, independentVarType, outcome)
}
\arguments{
\item{inputResults}{A list of data frames. Each object must include
model and processing results (output of ProcessResults()).}

\item{independentVarType}{The independent variable type ("gene" or "metabolite")}

\item{outcome}{The outcome type ("gene" or "metabolite")}
}
\description{
Construct a data frame that includes all information needed to build an igraph
object. This includes the names of the two co-regulated analytes, whether the
analytes have a positive or negative co-regulation, and the shape of the analytes.
}
