% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_BuildCoRegulationGraph.R
\name{ComputeCentrality}
\alias{ComputeCentrality}
\title{Return each analyte in the graph ranked by hub score. The analytes at the top
of the data frame are the hubs.}
\usage{
ComputeCentrality(coRegulationGraph, hubMetric)
}
\arguments{
\item{coRegulationGraph}{An igraph object. This graph is the co-regulation graph
generated using IntLIM analysis of analyte pairs.}

\item{hubMetric}{Can take one of four values:
1. "betweenness": the betweenness centrality (i.e. the percent of all paths in the
graph that pass through a node).
2. "kleinberg": the square of the adjacency matrix. This measures the connectivity
of the node.
3. "closeness": the average shortest path from a node to all other nodes.
4. "eccentricity": the maximum distance from a node to all other nodes.}
}
\description{
Return each analyte in the graph ranked by hub score. The analytes at the top
of the data frame are the hubs.
}
