% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_LearnGraphPredictionModel.R
\name{ComputeClassificationError}
\alias{ComputeClassificationError}
\title{Compute classification error.}
\usage{
ComputeClassificationError(true.Y, pred.Y)
}
\arguments{
\item{true.Y}{The true phenotype of each sample.}

\item{pred.Y}{The predicted phenotype of each sample.}
}
\description{
Compute classification error.
}
