% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientfunctions.R
\name{computeGradientSingleLayer}
\alias{computeGradientSingleLayer}
\title{Compute the gradient for a single layer neural network with the output
transformed by a sigmoid.}
\usage{
computeGradientSingleLayer(modelResults)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}
}
\description{
Compute the gradient for a single layer neural network with the output
transformed by a sigmoid.
}
