% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientfunctions.R
\name{computeHessianSingleLayer}
\alias{computeHessianSingleLayer}
\title{Compute the Hessian for a single layer neural network.
Note that this function is currently not implemented for categorical outcomes.}
\usage{
computeHessianSingleLayer(modelResults, convolution, pooling)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}

\item{convolution}{Whether or not to perform convolution}

\item{pooling}{Whether or not to perform pooling}
}
\description{
Compute the Hessian for a single layer neural network.
Note that this function is currently not implemented for categorical outcomes.
}
