% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_LearnGraphPredictionModel.R
\name{ComputeNRMSE}
\alias{ComputeNRMSE}
\title{Compute the normalized root mean squared error.}
\usage{
ComputeNRMSE(true.Y, pred.Y)
}
\arguments{
\item{true.Y}{The true phenotype of each sample.}

\item{pred.Y}{The predicted phenotype of each sample.}
}
\description{
Compute the normalized root mean squared error.
}
