% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{ConsolidateMultipleModelTypes}
\alias{ConsolidateMultipleModelTypes}
\title{If IntLIM is run on multiple types of pairs (e.g.
gene-gene and gene-metabolite) and on multiple folds,
the results will be arranged first by fold, then by type.
This function rearranges the results so that all model
types are concatenated under one fold.}
\usage{
ConsolidateMultipleModelTypes(results)
}
\arguments{
\item{results}{A list of lists of IntLIMResult data frames.}
}
\value{
A named list of IntLIMResult data frames.
}
\description{
If IntLIM is run on multiple types of pairs (e.g.
gene-gene and gene-metabolite) and on multiple folds,
the results will be arranged first by fold, then by type.
This function rearranges the results so that all model
types are concatenated under one fold.
}
