% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{ConsolidateMultipleModelTypesList}
\alias{ConsolidateMultipleModelTypesList}
\title{If IntLIM is run on multiple types of pairs (e.g.
gene-gene and gene-metabolite) and on multiple folds,
the results will be arranged first by fold, then by type.
This function rearranges the results so that all model
types are listed under one fold.}
\usage{
ConsolidateMultipleModelTypesList(results)
}
\arguments{
\item{results}{A list of lists of IntLIMResults.}
}
\value{
A named list of IntLIMResults.
}
\description{
If IntLIM is run on multiple types of pairs (e.g.
gene-gene and gene-metabolite) and on multiple folds,
the results will be arranged first by fold, then by type.
This function rearranges the results so that all model
types are listed under one fold.
}
