% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{CorrHeatmap_pdf}
\alias{CorrHeatmap_pdf}
\title{Plot correlation heatmap in pdf}
\usage{
CorrHeatmap_pdf(inputResults, viewer = T, top_pairs = 1200,
  treecuts = 2, palette = NULL)
}
\arguments{
\item{inputResults}{IntLimResults object (output of ProcessResults())}

\item{viewer}{whether the plot should be displayed in the RStudio viewer (T) or
in Shiny/Knittr (F)}

\item{top_pairs}{cutoff of the top pairs, sorted by adjusted p-values, to be plotted (plotting more than 1200 can take some time) (default: 1200)}

\item{treecuts}{number of clusters (of gene-metabolite pairs) to cut the tree into for color-coding}

\item{palette}{choose an RColorBrewer palette ("Set1", "Set2", "Set3",
"Pastel1", "Pastel2", "Paired", etc.) or submit a vector of colors}
}
\value{
a highcharter object
}
\description{
Plot correlation heatmap in pdf
}
