% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ReadData.R
\name{CreateCrossValFolds}
\alias{CreateCrossValFolds}
\title{Filter input data by abundance values (gene and metabolite data) and number of missing values (metabolite data only).}
\usage{
CreateCrossValFolds(
  inputFile,
  metabid = NULL,
  geneid = NULL,
  logmetab = FALSE,
  loggene = FALSE,
  folds
)
}
\arguments{
\item{inputFile}{input file in CSV format (see Despcription)}

\item{metabid}{name of column from metabolite meta data to be used as id
(required if a metabolite meta dadta file is present, must match metabolite abundances data)}

\item{geneid}{name of column from gene meta data to be used as id
(required if a gene meta data file is present, must match gene expression data)}

\item{logmetab}{whether or not to log metabolite values (T/F)}

\item{loggene}{whether or not to log gene values (T/F)}

\item{folds}{number of folds to create}
}
\value{
A set of MultiDataSet training and testing sets, of the following format:
list(list("train" = MultiDataSet, "test" = MultiDataSet), ... list("train" = MultiDataSet,
"test" = MultiDataSet))
}
\description{
Filter data by abundance (with user-input percentile cutoff) of missing values (with user-input percent cutoff). Missing values are commonly found in metabolomics data so the parameter currently only applies to metabolomics data.
}
