% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientfunctions.R
\name{CreateFilter}
\alias{CreateFilter}
\title{Create the initial filter to reflect the pooling operation. This function is
for scenarios where the weights are applied after pooling. Note that min 
and max pooling do not make sense in this scenario.}
\usage{
CreateFilter(poolingFilter, A.hat, X)
}
\arguments{
\item{poolingFilter}{A PoolingFilter object.}

\item{A.hat}{Convolutional operator.}

\item{X}{Input matrix from last training iteration.}
}
\description{
Create the initial filter to reflect the pooling operation. This function is
for scenarios where the weights are applied after pooling. Note that min 
and max pooling do not make sense in this scenario.
}
