% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{CreateIntLimObject}
\alias{CreateIntLimObject}
\title{Generic function to create constructor of MultiDataSet gene object}
\usage{
CreateIntLimObject(genefdata, metabfdata, pdata, geneid, metabid,
  metabdata, genedata, logmetab = FALSE, loggene = FALSE)
}
\arguments{
\item{genefdata}{gene meta data}

\item{metabfdata}{metabolite meta data}

\item{pdata}{sample meta data}

\item{geneid}{name of column from metabolite meta data to be used as id 
(required if a gene meta data file is present, must match gene expression matrix))}

\item{metabid}{name of column from gene meta data to be used as id
(required if a metabolite meta data file is present, must match metabolite abundances matrix))}

\item{metabdata}{metabolite abundances (samples are in columns)}

\item{genedata}{gene expression (samples are in columns)}

\item{logmetab}{T/F}

\item{loggene}{T/F}
}
\description{
Generic function to create constructor of MultiDataSet gene object
}
