% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ReadData.R
\name{CreateIntLimObjectPieces}
\alias{CreateIntLimObjectPieces}
\title{This is the helper function for ReadData. It is also used in CreateCrossValFolds.}
\usage{
CreateIntLimObjectPieces(
  inputFile,
  metabid = NULL,
  geneid = NULL,
  logmetab = FALSE,
  loggene = FALSE,
  suppressWarnings = FALSE
)
}
\arguments{
\item{inputFile}{input file in CSV format (see Despcription)}

\item{metabid}{name of column from metabolite meta data to be used as id
(required if a metabolite meta dadta file is present, must match metabolite abundances data)}

\item{geneid}{name of column from gene meta data to be used as id
(required if a gene meta data file is present, must match gene expression data)}

\item{logmetab}{whether or not to log metabolite values (T/F)}

\item{loggene}{whether or not to log gene values (T/F)}

\item{suppressWarnings}{whether to suppress warnings}
}
\value{
Named list of components for MultiDataSet
}
\description{
This is the helper function for ReadData. It is also used in CreateCrossValFolds.
}
