% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_LearnGraphPredictionModel.R
\name{CreateLineGraph}
\alias{CreateLineGraph}
\title{Create a line graph given the original, unweighted graph. In a line graph,
edges are nodes, and edges connected by a node are edges.}
\usage{
CreateLineGraph(predictionsByEdge, graphWithPredictions)
}
\arguments{
\item{predictionsByEdge}{Prediction levels corresponding to each edge.}

\item{graphWithPredictions}{Original graph data frame.}
}
\description{
Create a line graph given the original, unweighted graph. In a line graph,
edges are nodes, and edges connected by a node are edges.
}
