% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_LearnGraphPredictionModel.R
\name{CreatePoolingFilterKMeans}
\alias{CreatePoolingFilterKMeans}
\title{Create the graph pooling filter, given the adjacency matrix of the input graph.}
\usage{
CreatePoolingFilterKMeans(modelInputs, k, poolType)
}
\arguments{
\item{modelInputs}{An object of type "ModelInputs".}

\item{k}{The output dimensionality of the filter.}

\item{poolType}{One of "mean", "median", "max", or "min".}
}
\description{
Create the graph pooling filter, given the adjacency matrix of the input graph.
}
