% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_ProcessResults.R
\name{CreateResultsTable}
\alias{CreateResultsTable}
\title{Create results table, which includes significant gene:metabolite pairs, associated p-values, 
and correlations in each category evaluated.}
\usage{
CreateResultsTable(inputResults)
}
\arguments{
\item{inputResults}{IntLimResults object with model results (output of ProcessResults())}
}
\description{
Create results table, which includes significant gene:metabolite pairs, associated p-values, 
and correlations in each category evaluated.
}
\examples{
\dontrun{
dir <- system.file("extdata", package="IntLim", mustWork=TRUE)
csvfile <- file.path(dir, "NCItestinput.csv")
mydata <- ReadData(csvfile,metabid='id',geneid='id')
myres <- RunIntLim(mydata,stype="PBO_vs_Leukemia")
myres <- ProcessResults(myres,mydata)
mytable <- CreateResultsTable(myres)
}
}
