% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringfunctions.R
\name{doHierarchicalClustering}
\alias{doHierarchicalClustering}
\title{Perform hierarchical clustering of the unweighted graph.}
\usage{
doHierarchicalClustering(modelInputs)
}
\arguments{
\item{modelInputs}{An object of type "ModelInputs".}
}
\description{
Perform hierarchical clustering of the unweighted graph.
}
