% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_LearnGraphPredictionModel.R
\name{DoPrediction}
\alias{DoPrediction}
\title{Predict Y given current weights.}
\usage{
DoPrediction(modelResults, iteration, pooling, convolution)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}

\item{iteration}{The current iteration.}

\item{pooling}{Whether or not to pool the weights.}

\item{convolution}{Whether or not to perform convolution.}
}
\description{
Predict Y given current weights.
}
