% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_LearnGraphPredictionModel.R
\name{DoSingleTrainingIteration}
\alias{DoSingleTrainingIteration}
\title{Train the graph learning model, using the specifications in the ModelResults
class and storing the results in the ModelResults class.T}
\usage{
DoSingleTrainingIteration(
  modelResults,
  iteration,
  pooling,
  convolution,
  ridgeRegressionWeight,
  varianceWeight
)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}

\item{iteration}{The current iteration.}

\item{pooling}{Whether or not to pool the weights.}

\item{convolution}{Whether or not to perform convolution.}

\item{ridgeRegressionWeight}{The hyperparameter weight assigned
to the ridge regression parameter (often referred to as lambda in the
literature)}

\item{varianceWeight}{The hyperparameter weight assigned to the difference
in variances between Y and the predicted value of Y.}
}
\description{
Train the graph learning model, using the specifications in the ModelResults
class and storing the results in the ModelResults class.T
}
