% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FilterData.R
\name{FilterDataFolds}
\alias{FilterDataFolds}
\title{Filter input data by abundance values (gene and metabolite data) and number of missing values (metabolite data only).}
\usage{
FilterDataFolds(
  csvfile,
  stype,
  inputDataFolds,
  geneperc = 0,
  metabperc = 0,
  metabmiss = 0,
  class.covar = NULL,
  stype.type
)
}
\arguments{
\item{csvfile}{The path to the CSV file that contains the full dataset.}

\item{stype}{phenotype or outcome of interest.}

\item{inputDataFolds}{list of MultiDataSet objects (output of CreateCrossValFolds()) with gene expression, 
metabolite abundances, and associated meta-data}

\item{geneperc}{percentile cutoff (0-1) for filtering genes (e.g. remove genes with mean values 
< 'geneperc' percentile) (default: 0)}

\item{metabperc}{percentile cutoff (0-1) for filtering metabolites (default: no filtering of metabolites) (default:0)}

\item{metabmiss}{missing value percent cutoff (0-1) for filtering metabolites (metabolites with > 80\% missing values will be removed) (default:0)}

\item{class.covar}{class ("numeric" or "character") for each covariate. The following format
is required: list(covar1="numeric", covar2="character")}

\item{stype.type}{Either "factor" or "numeric". The outcome type.}
}
\value{
filtData list of MultiDataSet objects with input data after filtering
}
\description{
Filter data by abundance (with user-input percentile cutoff) of missing values (with user-input percent cutoff). Missing values are commonly found in metabolomics data so the parameter currently only applies to metabolomics data.
}
