% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_LearnGraphPredictionModel.R
\name{FindEdgesSharingNodes}
\alias{FindEdgesSharingNodes}
\title{Find edges that share nodes and add them to a data frame.}
\usage{
FindEdgesSharingNodes(
  predictionsByEdge,
  graphWithPredictions,
  nodeType1,
  nodeType2
)
}
\arguments{
\item{predictionsByEdge}{Prediction levels corresponding to each edge.}

\item{graphWithPredictions}{Original graph data frame.}

\item{nodeType1}{Either "to" or "from".}

\item{nodeType2}{Either "to or "from".}
}
\description{
Find edges that share nodes and add them to a data frame.
}
