% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringfunctions.R
\name{findHierarchicalClusteringVariance}
\alias{findHierarchicalClusteringVariance}
\title{Find the variance of each cluster in a hierarchical clustering.}
\usage{
findHierarchicalClusteringVariance(modelInputs)
}
\arguments{
\item{modelInputs}{An object of type "ModelInputs".}
}
\description{
Find the variance of each cluster in a hierarchical clustering.
}
