% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringfunctions.R
\name{findKClusters}
\alias{findKClusters}
\title{Find k clusters, optimizing based on the variance of the node
values in each cluster (averaged over all samples). This is a recursive function.}
\usage{
findKClusters(
  modelInputs,
  clusters,
  hClustResults,
  k,
  allClusters,
  allVariances
)
}
\arguments{
\item{modelInputs}{An object of the modelInputs class.}

\item{clusters}{A list of clusters, where each cluster is a set of nodes.}

\item{hClustResults}{The results of hierarchical clustering as performed using
the hclust() function.}

\item{k}{The number of clusters desired.}

\item{allClusters}{All clusters that have been explored.}

\item{allVariances}{All variances that have been computed.}
}
\description{
Find k clusters, optimizing based on the variance of the node
values in each cluster (averaged over all samples). This is a recursive function.
}
