% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_LearnGraphPredictionModel.R
\name{formatInput}
\alias{formatInput}
\title{Format the input for graph-based learning. This input consists of:
1. The Laplacian of a line graph built from the co-regulation graphs, where 
each node corresponds to a pair of analytes.
2. A prediction value for each node of the line graph, for each sample X.
3. The true prediction values Y for each sample X.}
\usage{
formatInput(
  predictionGraphs,
  coregulationGraph,
  inputData,
  stype,
  stype.class,
  edgeTypeList
)
}
\arguments{
\item{predictionGraphs}{A list of igraph objects, each of which includes
predictions for each edge.}

\item{coregulationGraph}{An igraph object containing the coregulation graph.}

\item{inputData}{MultiDataSet object (output of ReadData()) with gene expression,
metabolite abundances, and associated meta-data}

\item{stype}{The outcome}

\item{stype.class}{The class of the outcome ("numeric" or "categorical")}

\item{edgeTypeList}{List containing one or more of the following to include
in the line graph:
- "shared.outcome.analyte"
- "shared.independent.analyte"
- "analyte.chain"}
}
\description{
Format the input for graph-based learning. This input consists of:
1. The Laplacian of a line graph built from the co-regulation graphs, where 
each node corresponds to a pair of analytes.
2. A prediction value for each node of the line graph, for each sample X.
3. The true prediction values Y for each sample X.
}
