% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_LearnGraphPredictionModel.R
\name{formatInputAllFolds}
\alias{formatInputAllFolds}
\title{A wrapper for the formatInput class.}
\usage{
formatInputAllFolds(
  predictionGraphs,
  coregulationGraph,
  inputData,
  stype.class,
  edgeTypeList,
  testing = FALSE
)
}
\arguments{
\item{predictionGraphs}{A list of igraph objects, each of which includes
predictions for each edge.}

\item{coregulationGraph}{An igraph object containing the coregulation graph.}

\item{inputData}{List of MultiDataSet objects (output of CreateCrossValFolds()) 
with gene expression,
metabolite abundances, and associated meta-data}

\item{stype.class}{The class of the outcome ("numeric" or "categorical")}

\item{edgeTypeList}{List containing one or more of the following to include
in the line graph:
- "shared.outcome.analyte"
- "shared.independent.analyte"
- "analyte.chain"}

\item{testing}{A boolean indicating whether the testing data is to be used
(as opposed to the training data). Default is FALSE.}
}
\description{
A wrapper for the formatInput class.
}
