% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{formatSingleOmicInput}
\alias{formatSingleOmicInput}
\title{Function that returns a list of all data for samples in metabolite or gene dataset.}
\usage{
formatSingleOmicInput(
  inputData,
  stype = NULL,
  covar = NULL,
  class.covar = NULL,
  type = NULL
)
}
\arguments{
\item{inputData}{MultiDataSet object (output of ReadData())}

\item{stype}{category to color-code by (can be more than two categories)}

\item{covar}{vector of additional variables to be incorporated into model}

\item{class.covar}{class of additional variables}

\item{type}{either "metabolite" or "expression"}
}
\description{
Function that returns a list of all data for samples in metabolite or gene dataset.
}
