% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{GetCorrClusters}
\alias{GetCorrClusters}
\title{Returns the clusters found using CorrHeatmap.}
\usage{
GetCorrClusters(inputResults, treecuts = 2)
}
\arguments{
\item{inputResults}{Data frame (output of ProcessResults())}

\item{treecuts}{number of clusters (of pairs) to cut the tree 
into for color-coding}
}
\value{
A data frame including the independent and outcome analytes in each
for each pair and the cluster to which that pair belongs.
}
\description{
Returns the clusters found using CorrHeatmap.
}
