% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_LearnGraphPredictionModel.R
\name{GetErrorDeltasAllFolds}
\alias{GetErrorDeltasAllFolds}
\title{Run a prediction on new data using the graph learning model, and compute
the absolute error delta values.}
\usage{
GetErrorDeltasAllFolds(
  weights,
  modelInput,
  convolutions,
  minimum = NULL,
  maximum = NULL
)
}
\arguments{
\item{weights}{The list of all learned weights.}

\item{modelInput}{A list of objects of the ModelInput class.}

\item{convolutions}{Number of convolutions to perform. This should be the
same number of convolutions used in learning the weights.}

\item{minimum}{The minimum prediction value to allow. Default is NULL.}

\item{maximum}{The maximum prediction value to allow. Default is NULL.}
}
\description{
Run a prediction on new data using the graph learning model, and compute
the absolute error delta values.
}
