% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramplinkedinfunctions.R
\name{getGeneList}
\alias{getGeneList}
\title{Obtain lists of source IDs for inputting into RaMP or other pathway analysis
program
Requires sourceIDs and IDtypes
Otherwise will output names given
The IDtype should be as below}
\usage{
getGeneList(inputResults, inputData, outputGene = "id")
}
\arguments{
\item{inputResults}{results of IntLIM analysis}

\item{inputData}{IntLIM dataset containing feature data}

\item{outputGene}{format of genes.  Either names or a list of all
gene ids.  'id' means source ID, 'name' means gene entry names,
'mapping' provides names and source IDs}
}
\value{
a list of source IDs or names for genes of interests
}
\description{
+--------------------+
| IDtype             |
 +--------------------+
 | enzymeNomenclature |
 | ensembl            |
 | entrez             |
 | hmdb               |
 | kegg               |
 | uniprot            |
 +--------------------+
}
