% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramplinkedinfunctions.R
\name{getGeneRes}
\alias{getGeneRes}
\title{Obtain lists of source IDs for inputting into RaMP or other pathway analysis
program
Requires sourceID for gene suitable for RaMP
Otherwise will output names given}
\usage{
getGeneRes(inputResults, inputData, outputGene = "id")
}
\arguments{
\item{inputResults}{results of IntLIM analysis}

\item{inputData}{IntLIM dataset containing feature data}

\item{outputGene}{format of metabolites.  Either names or a list of all
gene ids.  'id' means source ID, 'name' means gene entry names,
'mapping' provides names and source IDs}
}
\value{
a list of source IDs or names for genes of interests
}
\description{
Obtain lists of source IDs for inputting into RaMP or other pathway analysis
program
Requires sourceID for gene suitable for RaMP
Otherwise will output names given
}
