% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramplinkedinfunctions.R
\name{getMetabRes}
\alias{getMetabRes}
\title{Obtain lists of source IDs for inputting into RaMP or other pathway analysis
program
Requires KEGG ids and/or HMDB ids
Otherwise will output names given}
\usage{
getMetabRes(inputResults, inputData, outputMetab = "id")
}
\arguments{
\item{inputResults}{results of IntLIM analysis}

\item{inputData}{IntLIM dataset containing feature data}

\item{outputMetab}{format of metabolites.  Either names or a list of all
metabolite ids.  'id' means source ID, 'name' means metabolite entry names,
'mapping' provides names and source IDs}
}
\value{
a list of source IDs or names for metabolites of interests
}
\description{
Obtain lists of source IDs for inputting into RaMP or other pathway analysis
program
Requires KEGG ids and/or HMDB ids
Otherwise will output names given
}
