% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{getQuantileForCoefficient}
\alias{getQuantileForCoefficient}
\title{Function that gets numeric cutoffs from percentile}
\usage{
getQuantileForCoefficient(tofilter, interactionCoeffPercentile)
}
\arguments{
\item{tofilter}{dataframe for percentile filtering}

\item{interactionCoeffPercentile}{percentile cutoff for absolute value of interaction coefficient}
}
\value{
vector with numeric cutoffs
}
\description{
Function that gets numeric cutoffs from percentile
}
