% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{getQuantileForInteractionCoefficient}
\alias{getQuantileForInteractionCoefficient}
\title{Function that gets numeric cutoffs from percentile}
\usage{
getQuantileForInteractionCoefficient(tofilter, interactionCoeffPercentile)
}
\arguments{
\item{tofilter}{dataframe for percentile filtering}

\item{interactionCoeffPercentile}{percentile cutoff for interaction coefficient (default bottom 10 percent (high negative coefficients) and top 10 percent (high positive coefficients))}
}
\value{
vector with numeric cutoffs
}
\description{
Function that gets numeric cutoffs from percentile
}
