% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{getStatsAllLM}
\alias{getStatsAllLM}
\title{Function that runs Linear Models for all genes or metabolites}
\usage{
getStatsAllLM(
  outcome,
  independentVariable,
  gene,
  metab,
  type,
  covar,
  covarMatrix,
  continuous,
  save.covar.pvals,
  remove.tri = FALSE
)
}
\arguments{
\item{outcome}{'metabolite' or 'gene' must be set as outcome/independent variable}

\item{independentVariable}{'metabolite' or 'gene' must be set as outcome/independent variable}

\item{gene}{gene dataset (incommon$gene)}

\item{metab}{metabolite dataset (incommon$metab)}

\item{type}{vector of sample type (by default, it will be used in the interaction term).
Only 2 categories are currently supported.}

\item{covar}{vector of additional vectors to consider}

\item{covarMatrix}{covariate matrix in (incommon$covar_matrix)}

\item{continuous}{indicate whether data is discrete (FALSE) or continuous (TRUE)}

\item{save.covar.pvals}{boolean to indicate whether or not to save the p-values of all covariates,
which can be analyzed later but will also lengthen computation time.}

\item{remove.tri}{boolean to indicate whether or not to remove the metabolite-metabolite
or gene-gene pair with the highest p-value across two duplicate models (e.g. m1~m2 and m2~m1)}
}
\value{
list of matrices (interaction.pvalues, interaction.adj.pvalues, interaction.coefficients)
}
\description{
Function that runs Linear Models for all genes or metabolites
}
