% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{getStatsAllLMMetabolitePairs}
\alias{getStatsAllLMMetabolitePairs}
\title{Function that runs Linear Models for all pairs of metabolites}
\usage{
getStatsAllLMMetabolitePairs(
  metab,
  type,
  covar,
  covarMatrix,
  continuous,
  save.covar.pvals,
  global = FALSE
)
}
\arguments{
\item{metab}{metabolite dataset in incommon MultiDataSet object (incommon$metab)}

\item{type}{vector of sample type (by default, it will be used in the interaction term).}

\item{covar}{vector of additional vectors to consider}

\item{covarMatrix}{covariate matrix in incommon MultiDataSet object (incommon$covar_matrix)}

\item{continuous}{indicate whether data is discrete (FALSE) or continuous (TRUE)}

\item{save.covar.pvals}{boolean to indicate whether or not to save the p-values of all covariates,
which can be analyzed later but will also lengthen computation time.}

\item{global}{boolean to indicate whether or not to return the p-value and coefficient of the
global co-regulation (as opposed to the differential co-reguation from the interaction term).}
}
\value{
list of matrices (interaction.pvalues, interaction.adj.pvalues, interaction.coefficients)
}
\description{
Function that runs Linear Models for all pairs of metabolites
}
