% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{HistogramGGPairs}
\alias{HistogramGGPairs}
\title{histogram of gene-gene pairs
depending upon gene}
\usage{
HistogramGGPairs(inputResults, type = "outcome", breaks = 50)
}
\arguments{
\item{inputResults}{IntLimResults object with model results (output of RunIntLim() 
and ProcessResults())}

\item{type}{'outcome' or 'independent'.  'outcome' set as default}

\item{breaks}{Number of breaks selected for histogram}
}
\description{
histogram of gene-gene pairs
depending upon gene
}
\examples{
\dontrun{
dir <- system.file("extdata", package="IntLIM", mustWork=TRUE)
csvfile <- file.path(dir, "NCItestinput.csv")
mydata <- ReadData(csvfile,metabid='id',geneid='id')
myres <- RunIntLim(mydata,stype="PBO_vs_Leukemia")
myres <- ProcessResults(inputResults=myres,inputData=mydata)
HistogramGMPairs(inputResults=myres)
}
}
