% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{HistogramGMPairs}
\alias{HistogramGMPairs}
\title{histogram of gene-metabolite pairs
depending upon metabolite or gene}
\usage{
HistogramGMPairs(inputResults, type = "metabolite", breaks = 50)
}
\arguments{
\item{inputResults}{IntLimResults object with model results (output of RunIntLim() and ProcessResults())}

\item{type}{'metabolite' or 'gene'.  'metabolite' set as default}

\item{breaks}{Number of breaks selected for histogram}
}
\description{
histogram of gene-metabolite pairs
depending upon metabolite or gene
}
\examples{
\dontrun{
dir <- system.file("extdata", package="IntLIM", mustWork=TRUE)
csvfile <- file.path(dir, "NCItestinput.csv")
mydata <- ReadData(csvfile,metabid='id',geneid='id')
myres <- RunIntLim(mydata,stype="PBO_vs_Leukemia")
myres <- ProcessResults(inputResults=myres,inputData=mydata)
HistogramGMPairs(inputResults=myres)
}
}
