% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{HistogramPairs}
\alias{HistogramPairs}
\title{histogram of analyte pairs
depending upon independent or outcome analyte}
\usage{
HistogramPairs(inputResults, type = "outcome", breaks = 50)
}
\arguments{
\item{inputResults}{Data frame with model results (output of ProcessResults())}

\item{type}{'independent' or 'outcome'.  'outcome' set as default}

\item{breaks}{Number of breaks selected for histogram}
}
\value{
No return value, called for side effects
}
\description{
histogram of analyte pairs
depending upon independent or outcome analyte
}
