% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringfunctions.R
\name{initializeClusters}
\alias{initializeClusters}
\title{Initialize a list of clusters for the graph, where each node is included
in its own cluster.}
\usage{
initializeClusters(hierarchicalClustering)
}
\arguments{
\item{hierarchicalClustering}{A data frame of merges from the hierarchical clustering.
This data frame is output by the function hclust().}
}
\description{
Initialize a list of clusters for the graph, where each node is included
in its own cluster.
}
