% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_LearnGraphPredictionModel.R
\name{InitializeGraphLearningModel}
\alias{InitializeGraphLearningModel}
\title{Find edges that share nodes and add them to a data frame.}
\usage{
InitializeGraphLearningModel(
  modelInputs,
  poolingFilter,
  iterations,
  convergenceCutoff,
  outcomeType,
  learningRate,
  activationType,
  weightsAfterPooling,
  optimizationType
)
}
\arguments{
\item{modelInputs}{An object of the ModelInput class.}

\item{poolingFilter}{A matrix that pools convolution results.}

\item{iterations}{Maximum number of iterations.}

\item{convergenceCutoff}{Cutoff for convergence.}

\item{outcomeType}{One of either "character" or "numeric"}

\item{learningRate}{Learning rate to use during training}

\item{activationType}{Activation function. May be "softmax", "sigmoid", 
or "tanh".}

\item{weightsAfterPooling}{Whether to include the weights after the pooling}

\item{optimizationType}{Type of optimization. May be "BGD", "momentum",
"adagrad", or "adam".
operation (as opposed to before). Must be TRUE or FALSE.}
}
\description{
Find edges that share nodes and add them to a data frame.
}
