% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringfunctions.R
\name{initializeMergeDataFrame}
\alias{initializeMergeDataFrame}
\title{Initialize the data frame of merges to include individual cluster and parent
cluster information.}
\usage{
initializeMergeDataFrame(hierarchicalClustering)
}
\arguments{
\item{hierarchicalClustering}{A data frame of merges from the hierarchical clustering.
This data frame is output by the function hclust().}
}
\description{
Initialize the data frame of merges to include individual cluster and parent
cluster information.
}
