% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IntLimData-class}
\alias{IntLimData-class}
\title{IntLimData class}
\description{
IntLimData class
}
\section{Slots}{

\describe{
\item{\code{analyteType1}}{A matrix of abundance, expression, or other levels for
a specific type of analyte (e.g. protein abundance, metabolite abundance, or
gene expression)}

\item{\code{analyteType2}}{A second matrix of abundance, expression, or other levels for
a specific type of analyte (e.g. protein abundance, metabolite abundance, or
gene expression)}

\item{\code{analyteType1MetaData}}{A data frame of metadata for analyte type 1.}

\item{\code{analyteType2MetaData}}{A data frame of metadata for analyte type 2.}

\item{\code{sampleMetaData}}{A data frame of covariate values from the patient data.}
}}

