% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{MarginalEffectsGraphDataframe}
\alias{MarginalEffectsGraphDataframe}
\title{Creates a dataframe of the marginal effect of phenotype}
\usage{
MarginalEffectsGraphDataframe(
  inputResults,
  inputData,
  geneOfInterest,
  metaboliteOfInterest,
  continuous
)
}
\arguments{
\item{inputResults}{IntLimResults object with model results (output of RunIntLim())}

\item{inputData}{Named list (output of 
FilterData()) with gene expression, metabolite abundances, 
and associated meta-data}

\item{geneOfInterest}{gene in gene-metabolite pair}

\item{metaboliteOfInterest}{metabolite in gene-metabolite pair}

\item{continuous}{whether or not the outcome is continuous (TRUE or FALSE)}
}
\value{
dataframe for further analysis
}
\description{
Creates a dataframe of the marginal effect of phenotype
}
