% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{MarginalEffectsGraphDataframeGenePairs}
\alias{MarginalEffectsGraphDataframeGenePairs}
\title{Creates a dataframe of the marginal effect of phenotype}
\usage{
MarginalEffectsGraphDataframeGenePairs(
  inputResults,
  inputData,
  geneOfInterest1,
  geneOfInterest2
)
}
\arguments{
\item{inputResults}{IntLimResults object with model results (output of RunIntLim())}

\item{inputData}{MultiDataSet object (output of ReadData()) with gene expression,}

\item{geneOfInterest1}{outcome gene in gene pair}

\item{geneOfInterest2}{independent gene in gene pair}
}
\value{
dataframe for further analysis
}
\description{
Creates a dataframe of the marginal effect of phenotype
}
