% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{MetaboliteSet-class}
\alias{MetaboliteSet-class}
\title{MetaboliteSet class extenstion for MultiDataSet object}
\description{
This class is specific to metabolomics data and will then be stored into the class
\code{MultiDataSet}, which stores multiple dataset types (e.g. metabolite and gene levels
in this case.
}
\section{Slots}{

\describe{
\item{\code{eSet}}{List of eSet elements}
}}

