% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ModelInput-class}
\alias{ModelInput-class}
\title{ModelInput class}
\description{
ModelInput class
}
\section{Slots}{

\describe{
\item{\code{A.hat}}{The Laplacian of a line graph built from the co-regulation graphs, where 
each node corresponds to a pair of analytes.}

\item{\code{node.wise.prediction}}{graph of co-regulation results from IntLIM}

\item{\code{true.phenotypes}}{data frame of true phenotypes given analyte pairs
and clinical covariates}

\item{\code{outcome.type}}{"numeric" or "categorical"}

\item{\code{coregulation.graph}}{the original co-regulation graph for the input data.}

\item{\code{line.graph}}{the line graph of the input data.}
}}

