% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{multi.which}
\alias{multi.which}
\title{A which for multidimensional arrays.
Mark van der Loo 16.09.2011}
\usage{
multi.which(A)
}
\arguments{
\item{A}{Boolean function defined over a matrix}
}
\value{
vector with numeric cutoffs
}
\description{
A which for multidimensional arrays.
Mark van der Loo 16.09.2011
}
