% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_RunPairwisePrediction.R
\name{multiplyCovariate}
\alias{multiplyCovariate}
\title{Multiply a covariate with its learned coefficients. This is straightforward for
numeric covariates, but requires some conversion for categorical covariates.}
\usage{
multiplyCovariate(
  coefficients,
  covariates,
  coef_cov_name,
  cov_cov_name,
  level1
)
}
\arguments{
\item{coefficients}{The coefficients of the selected covariates.}

\item{covariates}{The covariate values.}

\item{coef_cov_name}{The covariate name in the coefficient data}

\item{cov_cov_name}{The covariate name in the covariate data}

\item{level1}{The level of the vactor which equals "1" in the one-hot encoding.}
}
\description{
Multiply a covariate with its learned coefficients. This is straightforward for
numeric covariates, but requires some conversion for categorical covariates.
}
