% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsfunctions.R
\name{OutputStats}
\alias{OutputStats}
\title{Get some stats after reading in data}
\usage{
OutputStats(IntLimObject)
}
\arguments{
\item{IntLimObject}{output of ReadData()}
}
\value{
data.frame with some # of samples, features, etc.
}
\description{
Get some stats after reading in data
}
\examples{
dir <- system.file("extdata", package="IntLim", mustWork=TRUE)
csvfile <- file.path(dir, "test.csv")
mydata <- ReadData(csvfile,metabid='id',geneid='id')
OutputStats(mydata)
}
