% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermuteIntLIM.R
\name{PermuteIntLIM}
\alias{PermuteIntLIM}
\title{Run permutations of the IntLIM code to search for random cross-omic associations in dataset}
\usage{
PermuteIntLIM(
  data,
  stype,
  outcome,
  independent.var.type,
  analyte.metadata,
  covar = NULL,
  save.covar.pvals = TRUE,
  continuous = TRUE,
  filter.cutoff = 0.2,
  pval.cutoff = 0.05,
  adj.pval.cutoff = 0.2,
  num.permutations = 1,
  seed = 1
)
}
\arguments{
\item{data}{IntLimData object (output of ReadData()) with analylte levels and
associated sample meta-data}

\item{stype}{column name that represents sample type (by default, it will be used
in the interaction term). Only 2 categories are currently supported.}

\item{outcome}{'1' or '2' must be set as outcome/independent variable
(default is '1')}

\item{independent.var.type}{'1' or '2' must be set as independent variable
(default is '1')}

\item{analyte.metadata}{boolean value to indicate whether the user has included analyte 
metadata in the analysis}

\item{covar}{Additional variables from the phenotypic data that be integrated into linear model}

\item{save.covar.pvals}{boolean to indicate whether or not to save the p-values of all covariates,
which can be analyzed later but will also lengthen computation time. The default is FALSE.}

\item{continuous}{boolean to indicate whether the data is continuous or discrete}

\item{filter.cutoff}{percentile cutoff based on variance for both analytes to remove lowest variances in
dataset (default = 0.20)}

\item{pval.cutoff}{unadjusted p-value cutoff for number of significant multi-omic pairs (default = 0.05)}

\item{adj.pval.cutoff}{FDR adjusted p-value cutoff for number of significant multi-omic 
pairs (default = 0.20)}

\item{num.permutations}{Number of permutations to be ran (default = 1)}

\item{seed}{set.seed paramter allowing for custom random number generation seeds}
}
\value{
List object with 1st slot populated with dataframe containing the R^2 values of the models, 
and number of significant pairs before and after p-value adjustment. The 2nd slot in the list contains 
a string vector of the IDs of the significant pairs
}
\description{
This function allows users to test different permutations of the metadata with their analytes to ensure
that any pairs being deemed significant by IntIM are not being suggested due to random chance, as is 
sometimes a problem in correlative associations.
}
\examples{
\dontrun{
permutedResults <- PermuteIntLIM(data = mydata, stype = "PBO_vs_Leukemia", outcome = 1, 
                                 independent.var.type = 2, analyte.metadata = TRUE,
                                 num.permutations = 100)
}
}
