% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotCoRegulationGraph}
\alias{PlotCoRegulationGraph}
\title{Plot the graph with positive associations colored blue and negative associations
colored red.}
\usage{
PlotCoRegulationGraph(
  graph,
  title,
  saveInFile = NULL,
  vertices = NULL,
  truncateTo = 4
)
}
\arguments{
\item{graph}{The co-regulation graph.}

\item{title}{Title of plot}

\item{saveInFile}{Location where the file should be saved. If NULL, then the output
is plotted without being saved. Default is NULL.}

\item{vertices}{List of vectors of vertices to plot. This is used if one
wishes to focus on a subset of vertices. If NULL, then all vertices are plotted.
Default is NULL.}

\item{truncateTo}{Vertex names are truncated to the first "truncateTo" characters.
Default is 4. If NULL, names are not truncated.}
}
\description{
Plot the graph with positive associations colored blue and negative associations
colored red.
}
