% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotFoldOverlapUpSet}
\alias{PlotFoldOverlapUpSet}
\title{Makes an UpSet plot showing the filtered pairs of analytes found in each fold.
This plot should only be made for cross-validation data.}
\usage{
PlotFoldOverlapUpSet(inputResults)
}
\arguments{
\item{inputResults}{List of outputs of ProcessResultsAllFolds(), each of which
is a list of IntLIMResults.}
}
\value{
an UpSet plot
}
\description{
Makes an UpSet plot showing the filtered pairs of analytes found in each fold.
This plot should only be made for cross-validation data.
}
