% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotGMPair}
\alias{PlotGMPair}
\title{scatter plot of select gene and metab}
\usage{
PlotGMPair(inputData, stype, geneName, metabName)
}
\arguments{
\item{inputData}{IntLimObject output of ReadData() or FilterData()}

\item{stype}{vector of color}

\item{geneName}{string of select geneName}

\item{metabName}{string of select metabName}
}
\value{
a highcharter object
}
\description{
scatter plot of select gene and metab
}
\examples{
\dontrun{
dir <- system.file("extdata", package="IntLim", mustWork=TRUE)
csvfile <- file.path(dir, "NCIinput.csv")
mydata <- ReadData(csvfile,metabid='id',geneid='id')
stype = inputData[["expression"]]$PBO_vs_Leukemia
PlotGMPair(inputData,stype,"PRPF8","(p-Hydroxyphenyl)lactic acid")

}
}
