% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotGraphPredictions}
\alias{PlotGraphPredictions}
\title{Plot the graph for each sample with edges colored according to prediction.
Include a color scale for the predictions.}
\usage{
PlotGraphPredictions(
  graph,
  inputData,
  saveInDir = NULL,
  vertices = NULL,
  truncateTo = 4,
  titleStart = NULL
)
}
\arguments{
\item{graph}{The graph with predictions projected onto it.}

\item{inputData}{Named list (output of 
FilterData()) with gene expression, metabolite abundances, 
and associated meta-data}

\item{saveInDir}{Directory where files should be saved. If NULL, then the output
is plotted without being saved. Default is NULL.}

\item{vertices}{List of vectors of vertices to plot. This is used if one
wishes to focus on a subset of vertices. If NULL, then all vertices are plotted.
Default is NULL.}

\item{truncateTo}{Vertex names are truncated to the first "truncateTo" characters.
Default is 4. If NULL, names are not truncated.}

\item{titleStart}{The beginning of the title. Default is NULL.}
}
\description{
Plot the graph for each sample with edges colored according to prediction.
Include a color scale for the predictions.
}
