% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotGraphWeights}
\alias{PlotGraphWeights}
\title{Plot the graph with edges colored by weight in the final outcome.}
\usage{
PlotGraphWeights(graph, results)
}
\arguments{
\item{graph}{The co-regulation graph.}

\item{results}{A modelResults object.}
}
\description{
Plot the graph with edges colored by weight in the final outcome.
}
