% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotGraphWeightsHeatmap}
\alias{PlotGraphWeightsHeatmap}
\title{Plot the graph as a heatmap with edges colored by weight in the final outcome.}
\usage{
PlotGraphWeightsHeatmap(graph, results)
}
\arguments{
\item{graph}{The co-regulation graph.}

\item{results}{A modelResults object.}
}
\description{
Plot the graph as a heatmap with edges colored by weight in the final outcome.
}
