% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotGraphWeightsHeatmapAllFolds}
\alias{PlotGraphWeightsHeatmapAllFolds}
\title{A wrapper for PlotGraphWeights that applies the function across all folds.}
\usage{
PlotGraphWeightsHeatmapAllFolds(inputResults, inputData)
}
\arguments{
\item{inputResults}{The IntLIM results (from RunIntLimAllFolds())}

\item{inputData}{The input data (from CreateCrossValFolds())}
}
\description{
A wrapper for PlotGraphWeights that applies the function across all folds.
}
