% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotLineGraph}
\alias{PlotLineGraph}
\title{Plot the line graph for each sample with nodes colored according to prediction.
Include a color scale for the predictions.}
\usage{
PlotLineGraph(
  modelInput,
  stype,
  saveInDir = NULL,
  truncateTo = 2,
  titleStart = NULL,
  weights = NULL,
  analytes = NULL
)
}
\arguments{
\item{modelInput}{A list of ModelInput objects.}

\item{stype}{Outcome / phenotype}

\item{saveInDir}{Directory where file should be saved. If NULL, then the output
is plotted without being saved. Default is NULL.}

\item{truncateTo}{Analyte names are truncated to the first "truncateTo" characters.
Default is 4. If NULL, names are not truncated.}

\item{titleStart}{The beginning of the title. Default is NULL.}

\item{weights}{The weight assigned to each node. This is encoded using opacity.
If NULL, all nodes are opaque. Default is NULL.}

\item{analytes}{analytes of interest. Default is NULL}
}
\description{
Plot the line graph for each sample with nodes colored according to prediction.
Include a color scale for the predictions.
}
