% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotLineGraphAllFolds}
\alias{PlotLineGraphAllFolds}
\title{Wrapper for PlotLineGraph.}
\usage{
PlotLineGraphAllFolds(
  modelInputs,
  stype,
  saveInDir = NULL,
  analytes = NULL,
  truncateTo = 2,
  weights = NULL
)
}
\arguments{
\item{modelInputs}{A list of ModelInput objects.}

\item{stype}{Outcome / phenotype}

\item{saveInDir}{Directory where file should be saved. If NULL, then the output
is plotted without being saved. Default is NULL.}

\item{analytes}{List of vectors of analytes to plot. This is used if one
wishes to focus on a subset of analytes If NULL, then all vertices are plotted.
Default is NULL.}

\item{truncateTo}{Analyte names are truncated to the first "truncateTo" characters.
Default is 2. If NULL, names are not truncated.}

\item{weights}{The list of weights assigned to each node. This is encoded using opacity.
If NULL, all nodes are opaque. Default is NULL.}
}
\description{
Wrapper for PlotLineGraph.
}
