% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotMGPair}
\alias{PlotMGPair}
\title{scatter plot of metabolite-gene pairs (based on user selection)}
\usage{
PlotMGPair(
  inputData,
  stype = NULL,
  metabName,
  geneName,
  palette = "Set1",
  viewer = T
)
}
\arguments{
\item{inputData}{IntLimObject output of ReadData() or FilterData()}

\item{stype}{category to color-code by}

\item{metabName}{string of select metabName}

\item{geneName}{string of select geneName}

\item{palette}{choose an RColorBrewer palette ("Set1", "Set2", "Set3",
"Pastel1", "Pastel2", "Paired", etc.) or submit a vector of colors}

\item{viewer}{whether the plot should be displayed in the RStudio viewer (T) or
in Shiny/Knittr (F)}
}
\value{
a highcharter object
}
\description{
scatter plot of metabolite-gene pairs (based on user selection)
}
\examples{
\dontrun{
dir <- system.file("extdata", package="IntLIM", mustWork=TRUE)
csvfile <- file.path(dir, "NCItestinput.csv")
mydata <- ReadData(csvfile,metabid='id',geneid='id')
PlotGMPair(mydata,stype="PBO_vs_Leukemia","DLG4","(p-Hydroxyphenyl)lactic acid")

}
}
