% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotMMPair}
\alias{PlotMMPair}
\title{scatter plot of metabolite-metabolite pairs (based on user selection)}
\usage{
PlotMMPair(
  inputData,
  stype,
  metab1Name,
  metab2Name,
  palette = "Set1",
  viewer = T
)
}
\arguments{
\item{inputData}{IntLimObject output of ReadData() or FilterData()}

\item{stype}{category to color-code by}

\item{metab1Name}{string of select metab1Name}

\item{metab2Name}{string of select metab2Name}

\item{palette}{choose an RColorBrewer palette ("Set1", "Set2", "Set3",
"Pastel1", "Pastel2", "Paired", etc.) or submit a vector of colors}

\item{viewer}{whether the plot should be displayed in the RStudio viewer (T) or
in Shiny/Knittr (F)}
}
\value{
a highcharter object
}
\description{
scatter plot of metabolite-metabolite pairs (based on user selection)
}
