% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotPair}
\alias{PlotPair}
\title{scatter plot of gene-metabolite pairs (based on user selection)}
\usage{
PlotPair(
  inputData,
  stype,
  outcome,
  independentVariable,
  independentAnalyteOfInterest,
  outcomeAnalyteOfInterest,
  palette = "Set1",
  viewer = T
)
}
\arguments{
\item{inputData}{IntLimObject output of ReadData() or FilterData()}

\item{stype}{category to color-code by}

\item{outcome}{'metabolite' or 'gene' must be set as outcome/independent variable}

\item{independentVariable}{'metabolite' or 'gene' must be set as outcome/independent variable}

\item{independentAnalyteOfInterest}{independent analyte in pair}

\item{outcomeAnalyteOfInterest}{outcome analyte in pair}

\item{palette}{choose an RColorBrewer palette ("Set1", "Set2", "Set3",
"Pastel1", "Pastel2", "Paired", etc.) or submit a vector of colors}

\item{viewer}{whether the plot should be displayed in the RStudio viewer (T) or
in Shiny/Knittr (F)}
}
\description{
scatter plot of gene-metabolite pairs (based on user selection)
}
