% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotPairFlat}
\alias{PlotPairFlat}
\title{scatter plot of pairs (based on user selection). This version does not use
highcharter and instead plots a base R plot.}
\usage{
PlotPairFlat(
  inputData,
  inputResults,
  outcome,
  independentVariable,
  independentAnalyteOfInterest,
  outcomeAnalyteOfInterest,
  palette = "Set1"
)
}
\arguments{
\item{inputData}{IntLimObject output of ReadData() or FilterData()}

\item{inputResults}{Data frame with model results (output of ProcessResults())}

\item{outcome}{'1' or '2' must be set as outcome/independent variable}

\item{independentVariable}{'1' or '2' must be set as outcome/independent variable}

\item{independentAnalyteOfInterest}{independent analyte in pair}

\item{outcomeAnalyteOfInterest}{outcome analyte in pair}

\item{palette}{choose an RColorBrewer palette ("Set1", "Set2", "Set3",
"Pastel1", "Pastel2", "Paired", etc.) or submit a vector of colors}
}
\value{
No return value, called for side effects
}
\description{
scatter plot of pairs (based on user selection). This version does not use
highcharter and instead plots a base R plot.
}
