% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotPairResiduals}
\alias{PlotPairResiduals}
\title{scatter plot of pairs (based on user selection)}
\usage{
PlotPairResiduals(
  inputData,
  inputResults,
  outcome,
  independentVariable,
  independentAnalyteOfInterest,
  outcomeAnalyteOfInterest
)
}
\arguments{
\item{inputData}{IntLimObject output of ReadData() or FilterData()}

\item{inputResults}{Object of type ModelResults.}

\item{outcome}{'1' or '2' must be set as outcome/independent variable}

\item{independentVariable}{'1' or '2' must be set as outcome/independent variable}

\item{independentAnalyteOfInterest}{independent analyte in pair}

\item{outcomeAnalyteOfInterest}{outcome analyte in pair}
}
\value{
No return value, called for side effects
}
\description{
scatter plot of pairs (based on user selection)
}
