% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotPCA}
\alias{PlotPCA}
\title{PCA plots of data for QC}
\usage{
PlotPCA(inputData, viewer = T, stype = NULL, common = T,
  palette = "Set1")
}
\arguments{
\item{inputData}{IntLimObject output of ReadData()}

\item{viewer}{whether the plot should be displayed in the RStudio viewer (T) or
in Shiny/Knittr (F)}

\item{stype}{category to color-code by (can be more than two categories)}

\item{common}{whether or not samples that are in common between the metabolite and gene expression datasets should be plotted (T/F); default is TRUE}

\item{palette}{choose an RColorBrewer palette ("Set1", "Set2", "Set3",
"Pastel1", "Pastel2", "Paired", etc.) or submit a vector of colors}
}
\value{
a highcharter object
}
\description{
PCA plots of data for QC
}
\examples{
dir <- system.file("extdata", package="IntLim", mustWork=TRUE)
csvfile <- file.path(dir, "NCItestinput.csv")
mydata <- ReadData(csvfile,metabid='id',geneid='id')
PlotPCA(mydata,stype = "PBO_vs_Leukemia")
}
