% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PoolingFilter-class}
\alias{PoolingFilter-class}
\title{PoolingFilter class}
\description{
PoolingFilter class
}
\section{Slots}{

\describe{
\item{\code{filter}}{The filter that maps each dimension of the input to one of k clusters.}

\item{\code{filter.type}}{One of "mean", "min", "median", or "max".}

\item{\code{cluster.sizes}}{A vector of cluster sizes (by number of dimensions mapping}

\item{\code{individual.filters}}{A list of filters corresponding to each individual sample.
to the cluster)}
}}

