% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_LearnGraphPredictionModel.R
\name{PredictTesting}
\alias{PredictTesting}
\title{Run a prediction on new data using the graph learning model.}
\usage{
PredictTesting(modelResults, pooling, convolution, testInput)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}

\item{pooling}{Whether or not to pool the weights.}

\item{convolution}{Whether or not to perform convolution.}

\item{testInput}{An object of the ModelInput class.}

\item{iteration}{The current iteration.}
}
\description{
Run a prediction on new data using the graph learning model.
}
