% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_ProcessResults.R
\name{ProcessResults}
\alias{ProcessResults}
\title{Retrieve significant pairs, based on adjusted p-values, interaction coefficient
percentile, and R^2 values.}
\usage{
ProcessResults(
  inputResults,
  inputData,
  pvalcutoff = 0.05,
  coeffPercentile = 0,
  rsquaredCutoff = 0,
  coefficient = "interaction"
)
}
\arguments{
\item{inputResults}{IntLimResults object with model results (output of RunIntLim())}

\item{inputData}{IntLimData object (output of ReadData()) with analyte levels
and associated meta-data}

\item{pvalcutoff}{cutoff of FDR-adjusted p-value for filtering (default 0.05)}

\item{coeffPercentile}{percentile cutoff for absolute value of coefficient}

\item{rsquaredCutoff}{cutoff for lowest r-squared value}

\item{coefficient}{Coefficient to filter by. Default is interaction coefficient.
Other options are "stype" (phenotype) and "analyte" (independent analyte).}
}
\value{
IntLimResults object with model results.
}
\description{
Retrieve significant pairs, based on adjusted p-values, interaction coefficient
percentile, and R^2 values.
}
