% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_RunPairwisePrediction.R
\name{ProjectPredictionsOntoGraphAllFolds}
\alias{ProjectPredictionsOntoGraphAllFolds}
\title{This is a wrapper for ProjectPredictionsOntoGraph for all folds of data.}
\usage{
ProjectPredictionsOntoGraphAllFolds(predictions_list, coRegulationGraphs)
}
\arguments{
\item{predictions_list}{A list of lists of matrices of predictions. For each matrix,
each signficant pair of analytes results in a prediction for each subject. The
list may consist of: gene-gene model predictions, metabolite-metabolite model
predictions, metabolite-gene model predictions, and/or gene-metabolite model
predictions.}

\item{coRegulationGraphs}{A list of igraph objects. This graph is the co-regulation graph
generated using IntLIM analysis of analyte pairs.}
}
\description{
This is a wrapper for ProjectPredictionsOntoGraph for all folds of data.
}
