% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{pvalCoefVolcano}
\alias{pvalCoefVolcano}
\title{'volcano' plot (difference in correlations vs p-values)
of all pairs}
\usage{
pvalCoefVolcano(
  inputResults,
  inputData,
  nrpoints = 10000,
  pvalcutoff = 0.05,
  coefPercentileCutoff = 0.9
)
}
\arguments{
\item{inputResults}{Data frame with model results (output of ProcessResults())}

\item{inputData}{Named list (output of 
FilterData()) with analyte levels 
and associated meta-data}

\item{nrpoints}{number of points to be plotted in lowest density areas (see 'smoothScatter' documentation for more detail)}

\item{pvalcutoff}{cutoff of FDR-adjusted p-value for filtering (default 0.05)}

\item{coefPercentileCutoff}{cutoff of interaction coefficient percentile.}
}
\value{
a smoothScatter plot
}
\description{
'volcano' plot (difference in correlations vs p-values)
of all pairs
}
