% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{pvalCorrVolcano}
\alias{pvalCorrVolcano}
\title{'volcano' plot (difference in correlations vs p-values)
of all gene-metabolite pairs}
\usage{
pvalCorrVolcano(
  inputResults,
  inputData,
  nrpoints = 10000,
  diffcorr = 0.5,
  pvalcutoff = 0.05
)
}
\arguments{
\item{inputResults}{Data frame with model results (output of ProcessResults())}

\item{inputData}{Named list (output of 
FilterData()) with gene expression, metabolite abundances, 
and associated meta-data}

\item{nrpoints}{number of points to be plotted in lowest density areas (see 'smoothScatter' documentation for more detail)}

\item{diffcorr}{cutoff of differences in correlations for filtering (default 0.5)}

\item{pvalcutoff}{cutoff of FDR-adjusted p-value for filtering (default 0.05)}
}
\value{
a smoothScatter plot
}
\description{
'volcano' plot (difference in correlations vs p-values)
of all gene-metabolite pairs
}
