% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PValueBoxPlots}
\alias{PValueBoxPlots}
\title{Visualize the distribution of unadjusted p-values for all covariates
from linear models using a bar chart.}
\usage{
PValueBoxPlots(IntLimResults)
}
\arguments{
\item{IntLimResults}{output of RunIntLim()}
}
\description{
Visualize the distribution of unadjusted p-values for all covariates
from linear models using a bar chart.
}
\examples{
\dontrun{
dir <- system.file("extdata", package="IntLIM", mustWork=TRUE)
csvfile <- file.path(dir, "NCItestinput.csv")
mydata <- ReadData(csvfile,metabid='id',geneid='id')
myres <- RunIntLim(mydata,stype="PBO_vs_Leukemia")
DistPvalues(myres)
}
}
