% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ReadData.R
\name{ReadData}
\alias{ReadData}
\title{Read in CSV file}
\usage{
ReadData(
  inputFile,
  analyteType1id = "id",
  analyteType2id = "id",
  logAnalyteType1 = FALSE,
  logAnalyteType2 = FALSE,
  class.feat = NULL,
  suppressWarnings = FALSE
)
}
\arguments{
\item{inputFile}{input file in CSV format (see Despcription)}

\item{analyteType1id}{name of column from Analyte Type 1 meta data to be used as id
(required if an Analyte Type 1 meta data file is present, 
must match Analyte Type 1 data)}

\item{analyteType2id}{name of column from Analyte Type 2 meta data to be used as id
(required if an Analyte Type 2 meta data file is present, 
must match Analyte Type 2 data)}

\item{logAnalyteType1}{whether or not to log values for Analyte Type 1(T/F)}

\item{logAnalyteType2}{whether or not to log values for Analyte Type 2(T/F)}

\item{class.feat}{class ("factor" or "numeric") for each covariate. The following format
is required: list(covar1="numeric", covar2="factor")}

\item{suppressWarnings}{whether or not to suppress warnings}
}
\value{
IntLimData object with input data
}
\description{
The metadata associated with data files to be analyzed in IntLim is supplied
as a CSV file with two columns and 6 rows: 
   type,filenames
   analyteType1,myfilename
   analyteType2,myfilename
   analyteType1MetaData,myfilename (optional)
   analyteType2MetaData,myfilename (optional)
   sampleMetaData,myfilename
}
\details{
Note that all files supplied in the CSV file, and the CSV file itself should 
be placed in the same folder.  The software assumes will automatically retrieve 
the file path of
the input files (based on location of CSV files).  
Note also that the input data files should be in a specific format:
analyteType1/2: rows are analytes, columns are samples
analyteType1/2MetaData: rows are analytes, features are columns
sampleMetaData: rows are samples, features are columns
In addition, the first column of the sampleMetaData file is assumed to be the sample id, 
and those sample ids should match the columns of analyteType1/2 (e.g. it is required
that all sample ids in the analyteType1/2 are also in the sampleMetaData).
}
