% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{RemovePlusInCovars}
\alias{RemovePlusInCovars}
\title{RemovePlusInCovars}
\usage{
RemovePlusInCovars(covar = c(), sampleDataColnames)
}
\arguments{
\item{covar}{vector of additional vectors to consider}

\item{sampleDataColnames}{vector of column names, which is a superset
of the covar vector.}
}
\value{
list containing two elements: new covariates and new column names.
}
\description{
RemovePlusInCovars
}
