% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_RunIntLim.R
\name{RunIntLim}
\alias{RunIntLim}
\title{Run linear models and retrieve relevant statistics}
\usage{
RunIntLim(
  inputData,
  stype = NULL,
  outcome = "metabolite",
  covar = NULL,
  class.covar = NULL,
  continuous = FALSE,
  metabolite.pairs = FALSE
)
}
\arguments{
\item{inputData}{MultiDataSet object (output of ReadData()) with gene expression,
metabolite abundances, and associated meta-data}

\item{stype}{column name that represents sample type (by default, it will be used
in the interaction term). Only 2 categories are currently supported.}

\item{outcome}{'metabolite' or 'gene' must be set as outcome/independent variable
(default is 'metabolite')}

\item{covar}{Additional variables from the phenotypic data that be integrated into linear model}

\item{class.covar}{Describing whether additional variables are 'numeric' or 'categorial'}

\item{continuous}{boolean to indicate whether the data is continuous or discrete}

\item{metabolite.pairs}{boolean to indicate whether to return metabolite-metabolite pairs (TRUE) or gene-metabolite pairs (FALSE)}
}
\value{
IntLimModel object with model results
}
\description{
Run linear models and retrieve relevant statistics
}
\examples{
dir <- system.file("extdata", package="IntLIM", mustWork=TRUE)
csvfile <- file.path(dir, "NCItestinput.csv")
mydata <- ReadData(csvfile,metabid='id',geneid='id')
\dontrun{
myres <- RunIntLim(mydata,stype="PBO_vs_Leukemia")
}
}
