% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runIntLIMApp}
\alias{runIntLIMApp}
\title{run shiny app}
\usage{
runIntLIMApp(port = "127.0.0.1")
}
\arguments{
\item{port}{set port}
}
\value{
No return value, called for side effects
}
\description{
run shiny app
}
