% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_LearnGraphPredictionModel.R
\name{RunLeastSquaresOptimizationAllFolds}
\alias{RunLeastSquaresOptimizationAllFolds}
\title{Wrapper for RunLeastSquaresOptimization.}
\usage{
RunLeastSquaresOptimizationAllFolds(
  modelInput,
  convolutions,
  maskPercentile = 1,
  nrmseCutoff = NULL,
  corrCutoff = NULL,
  ridgeLambda = 0
)
}
\arguments{
\item{modelInput}{A list of objects of the ModelInput class.}

\item{convolutions}{Number of convolutions to perform.}

\item{maskPercentile}{The sum of error deltas will be computed over all
predictors. The lowest maskPercentile of predictors will be used to ensure
that reliable predictors are selected. By default, this value is set to 1,
which means that all predictors will be used.}

\item{nrmseCutoff}{This parameter allows you to use only predictors that
have NRMSE below a given cutoff. If NULL, a cutoff is not used. Default is NULL.}

\item{corrCutoff}{This parameter allows you to use only predictors that
have correlation above a given cutoff. If NULL, a cutoff is not used. Default is NULL.}

\item{ridgeLambda}{The lambda value to be used in ridge regression. The
higher the value of lambda, the more the sum of weights is minimized.
This encourages smaller weights. Default is 0.}
}
\description{
Wrapper for RunLeastSquaresOptimization.
}
