% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{RunLM}
\alias{RunLM}
\title{Function that runs linear models and returns interaction p-values.}
\usage{
RunLM(inputData, outcome = "metabolite", type = NULL)
}
\arguments{
\item{inputData}{MultiDataSet object (output of ReadData()) with gene}

\item{outcome}{'metabolite' or 'gene' must be set as outcome/independent variable 
(default is 'metabolite')}

\item{type}{vector of sample type (by default, it will be used in the interaction term).
Only 2 categories are currently supported.}
}
\description{
Function that runs linear models and returns interaction p-values.
}
