% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{RunLMGenePairs}
\alias{RunLMGenePairs}
\title{Function that runs linear models and returns interaction p-values for all gene pairs.}
\usage{
RunLMGenePairs(
  incommon,
  type = NULL,
  covar = NULL,
  continuous = FALSE,
  save.covar.pvals,
  global = FALSE
)
}
\arguments{
\item{incommon}{MultiDataSet object (output of ReadData()) with gene}

\item{type}{vector of sample type (by default, it will be used in the interaction term).
Only 2 categories are currently supported.}

\item{covar}{vector of additional vectors to consider}

\item{continuous}{boolean to indicate whether the data is continuous or discrete}

\item{save.covar.pvals}{boolean to indicate whether or not to save the p-values of all covariates,
which can be analyzed later but will also lengthen computation time}

\item{global}{boolean to indicate whether to save global co-regulation terms
(rather than interaction terms).}
}
\description{
Function that runs linear models and returns interaction p-values for all gene pairs.
}
