% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{RunLMMetabolitePairs}
\alias{RunLMMetabolitePairs}
\title{Function that runs linear models and returns interaction p-values for all metabolite pairs.}
\usage{
RunLMMetabolitePairs(
  incommon,
  type = NULL,
  covar = NULL,
  continuous = FALSE,
  save.covar.pvals = FALSE,
  keep.highest.pval = FALSE
)
}
\arguments{
\item{incommon}{MultiDataSet object (output of ReadData()) with gene}

\item{type}{vector of sample type (by default, it will be used in the interaction term).
Only 2 categories are currently supported.}

\item{covar}{vector of additional vectors to consider}

\item{continuous}{boolean to indicate whether the data is continuous or discrete}

\item{save.covar.pvals}{boolean to indicate whether or not to save the p-values of all covariates,
which can be analyzed later but will also lengthen computation time.
(rather than interaction terms).}

\item{keep.highest.pval}{boolean to indicate whether or not to remove the metabolite-metabolite
or gene-gene pair with the highest p-value across two duplicate models (e.g. m1~m2 and m2~m1)}
}
\description{
Function that runs linear models and returns interaction p-values for all metabolite pairs.
}
