% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_RunPairwisePrediction.R
\name{RunPairwisePredictionAllFolds}
\alias{RunPairwisePredictionAllFolds}
\title{Given each significant pairwise model and the input data, predict the phenotype
for each sample. Recall that IntLIM models take the following form:
m ~ beta0 + beta1(g) + beta2(phenotype) + beta3(g:phenotype) + beta4...n(covariates)
Therefore, to predict phenotype given the betas learned by IntLIM, we use the
following model:
p ~ (m - (beta0 + beta1(g) + beta4...n(covariates)) / (beta2 + beta3(g))}
\usage{
RunPairwisePredictionAllFolds(
  inputResults,
  inputData,
  stype = NULL,
  covar = NULL,
  testing = FALSE,
  independentVarType,
  outcome
)
}
\arguments{
\item{inputResults}{A list of IntLimResults objects. Each object must include
model and processing results (output of ProcessResultsAllFolds()). All results must
include learned covariate weights (i.e. must be run with save.covar.pvals = TRUE)}

\item{inputData}{MultiDataSet object (output of CreateCrossValFolds()) with gene expression,
metabolite abundances, and associated meta-data}

\item{stype}{The phenotype (outcome) to predict. This can be either a categorical
or numeric outcome.}

\item{covar}{The clinical covariates to include in the model. These should be the same
covariates that were included when running the IntLIM linear models.}

\item{testing}{Boolean indicating whether this is testing data. FALSE by default.}

\item{independentVarType}{The independent variable type ("gene" or "metabolite")}

\item{outcome}{The outcome type ("gene" or "metabolite")}
}
\description{
Given each significant pairwise model and the input data, predict the phenotype
for each sample. Recall that IntLIM models take the following form:
m ~ beta0 + beta1(g) + beta2(phenotype) + beta3(g:phenotype) + beta4...n(covariates)
Therefore, to predict phenotype given the betas learned by IntLIM, we use the
following model:
p ~ (m - (beta0 + beta1(g) + beta4...n(covariates)) / (beta2 + beta3(g))
}
