% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{scatterPlot}
\alias{scatterPlot}
\title{scatter plot of select gene and metab}
\usage{
scatterPlot(data)
}
\arguments{
\item{data}{formatted data}
}
\value{
a highcharter object
}
\description{
scatter plot of select gene and metab
}
