% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientfunctions.R
\name{SigmoidWithCorrection}
\alias{SigmoidWithCorrection}
\title{Apply the Sigmoid function to a numeric vector input. Note that we expect inputs
to be between 1 and 2, not -1 and 1. We therefore correct accordingly.}
\usage{
SigmoidWithCorrection(input)
}
\arguments{
\item{input}{A vectorized numeric input.}
}
\description{
Apply the Sigmoid function to a numeric vector input. Note that we expect inputs
to be between 1 and 2, not -1 and 1. We therefore correct accordingly.
}
