% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientfunctions.R
\name{SoftmaxWithCorrection}
\alias{SoftmaxWithCorrection}
\title{Apply the Softmax function to a numeric vector input. Note that we expect inputs
to be between 1 and 2, not 0 and 1. We therefore correct accordingly by subtracting
1 from the input and adding 1 to the output.}
\usage{
SoftmaxWithCorrection(input)
}
\arguments{
\item{input}{A vectorized numeric input.}
}
\description{
Apply the Softmax function to a numeric vector input. Note that we expect inputs
to be between 1 and 2, not 0 and 1. We therefore correct accordingly by subtracting
1 from the input and adding 1 to the output.
}
