% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientfunctions.R
\name{BackpropagateSingleLayer}
\alias{BackpropagateSingleLayer}
\title{Run backpropagation for a single layer. In backpropagation, a gradient is
is computed by taking partial derivatives for each of the model weights. Given
a learning rate, the weights are adjusted according to the gradient.}
\usage{
BackpropagateSingleLayer(
  modelResults,
  iteration,
  convolution,
  pooling,
  ridgeRegressionWeight,
  varianceWeight
)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}

\item{iteration}{The current iteration.}

\item{convolution}{Whether or not to perform convolution}

\item{pooling}{Whether or not to perform pooling}

\item{ridgeRegressionWeight}{The hyperparameter weight assigned
to the ridge regression parameter (often referred to as lambda in the
literature)}

\item{varianceWeight}{The hyperparameter weight assigned to the difference
in variances between Y and the predicted value of Y.}
}
\description{
Run backpropagation for a single layer. In backpropagation, a gradient is
is computed by taking partial derivatives for each of the model weights. Given
a learning rate, the weights are adjusted according to the gradient.
}
