% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_BuildCoRegulationGraph.R
\name{BuildCoRegulationGraph}
\alias{BuildCoRegulationGraph}
\title{Construct a co-regulation graph where nodes are genes and metabolites, and edges
are present between nodes and edges that meet the processing cutoffs. Note that
the ProcessResults function must be run before BuildCoRegulationGraph is run.
This is because the BuildCoRegulationGraph uses the output from ProcessResults
to generate the graph. The graph generated by this function is undirected and
is an igraph object.}
\usage{
BuildCoRegulationGraph(inputResults, inputData)
}
\arguments{
\item{inputResults}{A list of IntLimResults objects. Each object must include
model and processing results (output of ProcessResults()).}

\item{inputData}{MultiDataSet object (output of ReadData()) with gene expression,
metabolite abundances, and associated meta-data}
}
\description{
Construct a co-regulation graph where nodes are genes and metabolites, and edges
are present between nodes and edges that meet the processing cutoffs. Note that
the ProcessResults function must be run before BuildCoRegulationGraph is run.
This is because the BuildCoRegulationGraph uses the output from ProcessResults
to generate the graph. The graph generated by this function is undirected and
is an igraph object.
}
