% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_BuildCoRegulationGraph.R
\name{BuildGraphDataFrame}
\alias{BuildGraphDataFrame}
\title{Construct a data frame that includes all information needed to build an igraph
object. This includes the names of the two co-regulated analytes, whether the
analytes have a positive or negative co-regulation, and the shape of the analytes.}
\usage{
BuildGraphDataFrame(inputResults, inputData)
}
\arguments{
\item{inputResults}{A list of IntLimResults objects. Each object must include
model and processing results (output of ProcessResults()).}

\item{inputData}{MultiDataSet object (output of ReadData()) with gene expression,
metabolite abundances, and associated meta-data}
}
\description{
Construct a data frame that includes all information needed to build an igraph
object. This includes the names of the two co-regulated analytes, whether the
analytes have a positive or negative co-regulation, and the shape of the analytes.
}
