% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientfunctions.R
\name{computeGradientSingleLayer}
\alias{computeGradientSingleLayer}
\title{Compute the gradient for a single layer neural network.}
\usage{
computeGradientSingleLayer(modelResults, convolution, pooling)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}

\item{convolution}{Whether or not to perform convolution}

\item{pooling}{Whether or not to perform pooling}
}
\description{
Compute the gradient for a single layer neural network.
}
