% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{CorrHeatmap}
\alias{CorrHeatmap}
\title{Plot correlation heatmap}
\usage{
CorrHeatmap(inputResults, viewer = T)
}
\arguments{
\item{inputResults}{IntLimResults object (output of ProcessResults())}

\item{viewer}{whether the plot should be displayed in the RStudio viewer (T) or
in Shiny/Knittr (F)}
}
\value{
a highcharter object
}
\description{
Plot correlation heatmap
}
\examples{
dir <- system.file("extdata", package="IntLim", mustWork=TRUE)
csvfile <- file.path(dir, "test.csv")
mydata <- ReadData(csvfile,metabid='id',geneid='id')
myres <- RunIntLim(mydata,stype="cancertype")
myres <- ProcessResults(myres,mydata)
CorrHeatmap(myres)
}
