% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{CorrHeatmap}
\alias{CorrHeatmap}
\title{Plot correlation heatmap}
\usage{
CorrHeatmap(inputResults, viewer = T, top_pairs = 1200, treecuts = 2)
}
\arguments{
\item{inputResults}{IntLimResults object (output of ProcessResults())}

\item{viewer}{whether the plot should be displayed in the RStudio viewer (T) or
in Shiny/Knittr (F)}

\item{top_pairs}{cutoff of the top pairs, sorted by adjusted p-values, to be plotted (plotting more than 1200 can take some time) (default: 1200)}

\item{treecuts}{number of clusters (of gene-metabolite pairs) to cut the tree into for color-coding}
}
\value{
a highcharter object
}
\description{
Plot correlation heatmap
}
\examples{
\dontrun{
dir <- system.file("extdata", package="IntLim", mustWork=TRUE)
csvfile <- file.path(dir, "NCItestinput.csv")
mydata <- ReadData(csvfile,metabid='id',geneid='id')
myres <- RunIntLim(mydata,stype="PBO_vs_Leukemia")
myres <- ProcessResults(myres,mydata)
CorrHeatmap(myres)
}
}
