% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{CorrHeatmap}
\alias{CorrHeatmap}
\title{Plot correlation heatmap}
\usage{
CorrHeatmap(
  inputResults,
  viewer = T,
  top_pairs = 1200,
  treecuts = 2,
  palette = NULL,
  static = FALSE,
  html.file = NULL,
  pdf.file = NULL
)
}
\arguments{
\item{inputResults}{Data frame (output of ProcessResults())}

\item{viewer}{whether the plot should be displayed in the RStudio viewer (T) or
in Shiny/Knittr (F)}

\item{top_pairs}{cutoff of the top pairs, sorted by adjusted p-values, to be plotted (plotting more than 1200 can take some time) (default: 1200)}

\item{treecuts}{number of clusters (of pairs) to cut the tree into for color-coding}

\item{palette}{choose an RColorBrewer palette ("Set1", "Set2", "Set3",
"Pastel1", "Pastel2", "Paired", etc.) or submit a vector of colors}

\item{static}{allows user to decide whether heatmap is interactive or static}

\item{html.file}{allows user to specify file path to output heatmap onto (used for non-static heatmaply objects)}

\item{pdf.file}{allows user to specify file path to output heatmap onto (used for static heatmap.2 objects)}
}
\value{
a highcharter object
}
\description{
Plot correlation heatmap
}
