% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{CreateResultsTable}
\alias{CreateResultsTable}
\title{Create results table}
\usage{
CreateResultsTable(inputResults)
}
\arguments{
\item{inputResults}{IntLimResults object with model results (output of ProcessResults())}
}
\description{
Create results table
}
