% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{DistPvalues}
\alias{DistPvalues}
\title{Visualize the distribution of unadjusted p-values from linear models}
\usage{
DistPvalues(IntLimResults)
}
\arguments{
\item{IntLimResults}{output of RunIntLim()}
}
\description{
Visualize the distribution of unadjusted p-values from linear models
}
\examples{
dir <- system.file("extdata", package="IntLim", mustWork=TRUE)
csvfile <- file.path(dir, "test.csv")
mydata <- ReadData(csvfile,metabid='id',geneid='id')
myres <- RunIntLim(mydata,stype="cancertype")
DistPvalues(myres)
}
