% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FilterData.R
\name{FilterData}
\alias{FilterData}
\title{Filter data by abundance and/or CV}
\usage{
FilterData(inputData, geneperc = 5, metabperc = 0)
}
\arguments{
\item{inputData}{MultiDataSet object (output of ReadData()) with gene expression, 
metabolite abundances, and associated meta-data}

\item{geneperc}{percentile cutoff for filtering genes (e.g. remove genes with mean values 
< 'geneperc' percentile) (default: 5th percentile)}

\item{metabperc}{percentile cutoff for filtering metabolites (default: no filtering of metabolites)}
}
\value{
filtData MultiDataSet object with input data after filtering
}
\description{
Filter data by abundance and/or CV
}
\examples{
dir <- system.file("extdata", package="IntLim", mustWork=TRUE)
csvfile <- file.path(dir, "NCItestinput.csv")
inputData <- ReadData(csvfile,metabid='id',geneid='id')
inputDatafilt <- FilterData(inputData,geneperc=5)
}
