% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FilterData.R
\name{FilterData}
\alias{FilterData}
\title{Filter data by abundance and/or CV}
\usage{
FilterData(inputData, geneCV = NULL, metabCV = NULL, geneperc = NULL,
  metabperc = NULL)
}
\arguments{
\item{inputData}{MultiDataSet object (output of ReadData()) with gene expression, 
metabolite abundances, and associated meta-data}

\item{geneCV}{CV cutoff to be used for filtering genes (e.g. remove genes CV < 5)}

\item{metabCV}{CV cutoff to be used for filtering metabolites}

\item{geneperc}{percentile cutoff for filtering genes (e.g. remove genes with values < 15th
percentile)}

\item{metabperc}{percentile cutoff for filtering metabolites}
}
\value{
filtData MultiDataSet object with input data after filtering
}
\description{
Filter data by abundance and/or CV
}
\examples{
dir <- system.file("extdata", package="IntLim", mustWork=TRUE)
csvfile <- file.path(dir, "test.csv")
mydata <- ReadData(csvfile,metabid='BIOCHEMICAL',geneid='X')
mydatafilt <- FilterData(mydata,geneperc=15,metabCV=10)
}
