% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FilterData.R
\name{FilterData}
\alias{FilterData}
\title{Filter input data by abundance values (gene and metabolite data) and number of missing values (metabolite data only).}
\usage{
FilterData(
  inputData,
  stype,
  geneperc = 0,
  metabperc = 0,
  metabmiss = 0,
  class.covar = NULL,
  imputedMetabCutoff = NULL,
  imputedSampleCutoff = NULL,
  cvCutoffGenes = NULL,
  cvCutoffMetabs = NULL,
  log2scaleGenes = NULL,
  log2scaleMetabs = NULL,
  zScaleGenes = FALSE,
  zScaleMetabs = FALSE,
  stype.type
)
}
\arguments{
\item{inputData}{MultiDataSet object (output of ReadData()) with gene expression, 
metabolite abundances, and associated meta-data}

\item{stype}{phenotype or outcome of interest.}

\item{geneperc}{percentile cutoff (0-1) for filtering genes (e.g. remove genes with mean values 
< 'geneperc' percentile) (default: 0)}

\item{metabperc}{percentile cutoff (0-1) for filtering metabolites (default: no filtering of metabolites) (default:0)}

\item{metabmiss}{missing value percent cutoff (0-1) for filtering metabolites (metabolites with > 80\% missing values will be removed) (default:0)}

\item{class.covar}{class ("numeric" or "character") for each covariate. The following format
is required: list(covar1="numeric", covar2="character")}

\item{imputedMetabCutoff}{Similar to the metabperc parameter, but for imputed values.
Checks for minimum imputation above a given percentile, and removes the metabolite if
the condition is met. Default is NULL (no filtering),}

\item{imputedSampleCutoff}{Checks how many metabolites have been imputed for a given
sample, and if above the threshold, the sample is removed. Default is NULL (no filtering).}

\item{cvCutoffGenes}{Checks for coefficient of variation below a given threshold. Genes
that exceed the cutoff are removed.}

\item{cvCutoffMetabs}{Checks for coefficient of variation below a given threshold. Metabolites
that exceed the cutoff are removed.}

\item{log2scaleGenes}{Whether or not to log-scale the genes. Default is FALSE.}

\item{log2scaleMetabs}{Whether or not to log-scale the metabolites. Default is FALSE.}

\item{zScaleGenes}{Whether or not to z-scale each gene (subtract mean and divide
by standard deviation). This brings all genes onto the same scale. Default is FALSE.}

\item{zScaleMetabs}{Whether or not to z-scale each metabolite (subtract mean and divide
by standard deviation). This brings all metabolites onto the same scale. Default is FALSE.}

\item{stype.type}{Either "factor" or "numeric". The outcome type.}
}
\value{
filtData MultiDataSet object with input data after filtering
}
\description{
Filter data by abundance (with user-input percentile cutoff) of missing values (with user-input percent cutoff). Missing values are commonly found in metabolomics data so the parameter currently only applies to metabolomics data.
}
\examples{
dir <- system.file("extdata", package="IntLIM", mustWork=TRUE)
csvfile <- file.path(dir, "NCItestinput.csv")
inputData <- ReadData(csvfile,metabid='id',geneid='id')
inputDatafilt <- FilterData(inputData = inputData,stype = "PBO_vs_Leukemia", 
geneperc=0.5, stype.type = "factor")
}
