% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalfunctions.R
\name{FilterFromKeepers}
\alias{FilterFromKeepers}
\title{Filter a data set given the analytes to keep.}
\usage{
FilterFromKeepers(inputData, type1keepers, type2keepers)
}
\arguments{
\item{inputData}{IntLimData object (output of ReadData()) with analyte leves
and associated meta-data}

\item{type1keepers}{Type 1 analytes to keep.}

\item{type2keepers}{Type 2 analytes to keep.}
}
\value{
filtData IntLimData object with input data after filtering
}
\description{
Filter a data set given the analytes to keep.
}
