% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FilterData.R
\name{FilterFromKeepers}
\alias{FilterFromKeepers}
\title{Filter a data set given the metabolites and genes to keep.}
\usage{
FilterFromKeepers(inputData, genekeepers, metabkeepers)
}
\arguments{
\item{inputData}{MultiDataSet object (output of ReadData()) with gene expression, 
metabolite abundances, and associated meta-data}

\item{genekeepers}{Genes to keep.}

\item{metabkeepers}{Metabolites to keep.}
}
\value{
filtData MultiDataSet object with input data after filtering
}
\description{
Filter a data set given the metabolites and genes to keep.
}
