% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{formatSingleOmicInput}
\alias{formatSingleOmicInput}
\title{Function that returns a list of all data for samples in metabolite or gene dataset.}
\usage{
formatSingleOmicInput(inputData, stype = NULL, type = NULL, continuous = FALSE)
}
\arguments{
\item{inputData}{MultiDataSet object (output of ReadData())}

\item{stype}{category to color-code by (can be more than two categories)}

\item{type}{either "metabolite" or "expression"}

\item{continuous}{boolean value indicating whether outcome is continuous}
}
\description{
Function that returns a list of all data for samples in metabolite or gene dataset.
}
