% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{getCommon}
\alias{getCommon}
\title{Function that returns a list of all data for samples in common between metabolite and gene datasets}
\usage{
getCommon(inputData, stype = NULL, addvar = NULL, class.addvar = NULL)
}
\arguments{
\item{inputData}{MultiDataSet object (output of ReadData())}

\item{stype}{category to color-code by (can be more than two categories)}

\item{addvar}{vector of additional variables to be incorporated into model}

\item{class.addvar}{class of additional variables}
}
\description{
Function that returns a list of all data for samples in common between metabolite and gene datasets
}
