% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{getStatsAllLM}
\alias{getStatsAllLM}
\title{Function that runs Linear Models for all genes or metabolites}
\usage{
getStatsAllLM(outcome, gene, metab, type, covar, covarMatrix, continuous)
}
\arguments{
\item{outcome}{'metabolite' or 'gene' must be set as outcome/independent variable}

\item{gene}{gene dataset in incommon MultiDataSet object (incommon$gene)}

\item{metab}{metabolite dataset in incommon MultiDataSet object (incommon$metab)}

\item{type}{vector of sample type (by default, it will be used in the interaction term).
Only 2 categories are currently supported.}

\item{covar}{vector of additional vectors to consider}

\item{covarMatrix}{covariate matrix in incommon MultiDataSet object (incommon$covar_matrix)}

\item{continuous}{indicate whether data is discrete (FALSE) or continuous (TRUE)}
}
\value{
list of matrices (interaction.pvalues, interaction.adj.pvalues, interaction.coefficients)
}
\description{
Function that runs Linear Models for all genes or metabolites
}
