% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{getStatsAllLMGenePairs}
\alias{getStatsAllLMGenePairs}
\title{Function that runs Linear Models for all pairs of genes}
\usage{
getStatsAllLMGenePairs(
  gene,
  type,
  covar,
  covarMatrix,
  continuous,
  save.covar.pvals,
  remove.tri = FALSE
)
}
\arguments{
\item{gene}{metabolite dataset in incommon MultiDataSet object (incommon$gene)}

\item{type}{vector of sample type (by default, it will be used in the interaction term).}

\item{covar}{vector of additional vectors to consider}

\item{covarMatrix}{covariate matrix in incommon MultiDataSet object (incommon$covar_matrix)}

\item{continuous}{indicate whether data is discrete (FALSE) or continuous (TRUE)}

\item{save.covar.pvals}{boolean to indicate whether or not to save the p-values of all covariates,
which can be analyzed later but will also lengthen computation time.}

\item{remove.tri}{boolean to indicate whether or not to remove the metabolite-metabolite
or gene-gene pair with the highest p-value across two duplicate models (e.g. m1~m2 and m2~m1)}
}
\value{
list of matrices (interaction.pvalues, interaction.adj.pvalues, interaction.coefficients)
}
\description{
Function that runs Linear Models for all pairs of genes
}
