% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{getStatsAllLMMetabolitePairs}
\alias{getStatsAllLMMetabolitePairs}
\title{Function that runs Linear Models for all pairs of metabolites}
\usage{
getStatsAllLMMetabolitePairs(
  outcome,
  metab,
  type,
  covar,
  covarMatrix,
  continuous
)
}
\arguments{
\item{outcome}{'metabolite' must be set as outcome/independent variable}

\item{metab}{metabolite dataset in incommon MultiDataSet object (incommon$metab)}

\item{type}{vector of sample type (by default, it will be used in the interaction term).}

\item{covar}{vector of additional vectors to consider}

\item{covarMatrix}{covariate matrix in incommon MultiDataSet object (incommon$covar_matrix)}

\item{continuous}{indicate whether data is discrete (FALSE) or continuous (TRUE)}
}
\value{
list of matrices (interaction.pvalues, interaction.adj.pvalues, interaction.coefficients)
}
\description{
Function that runs Linear Models for all pairs of metabolites
}
