% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{getstatsOneLM}
\alias{getstatsOneLM}
\title{Function that runs linear models for one gene vs all metabolites}
\usage{
getstatsOneLM(form, clindata, arraydata)
}
\arguments{
\item{form}{LM formulat (typically m~g+t+g:t)}

\item{arraydata}{matrix of metabolite values}

\item{clindatta}{data frame with 1st column: expression of one analyte; 2nd column
sample type (e.g. cancer/non-cancer)}
}
\description{
Function that runs linear models for one gene vs all metabolites
}
