% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{getstatsOneLM}
\alias{getstatsOneLM}
\title{Function that runs linear models for analyte vs. all analytes of the other type}
\usage{
getstatsOneLM(form, clindata, arraydata, analytename, suppressWarnings = FALSE)
}
\arguments{
\item{form}{LM formulat (typically m~g+t+g:t)}

\item{clindata}{data frame with 1st column: expression of one analyte; 2nd column
sample type (e.g. cancer/non-cancer)}

\item{arraydata}{matrix of analyte values}

\item{analytename}{name of independent analyte in the model}

\item{suppressWarnings}{whether or not to suppress warnings}
}
\description{
Function that runs linear models for analyte vs. all analytes of the other type
}
