% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{InteractionCoefficientGraph}
\alias{InteractionCoefficientGraph}
\title{Graphs a scatterplot of gene-metabolite pairs vs. the interaction coefficient for the gene-metabolite pair}
\usage{
InteractionCoefficientGraph(inputResults,
  interactionCoeffPercentile = 0.1)
}
\arguments{
\item{inputResults}{IntLimResults object with model results (output of RunIntLim())}

\item{interactionCoeffPercentile}{percentile cutoff for interaction coefficient (default bottom 10 percent (high negative coefficients) and top 10 percent (high positive coefficients))}
}
\value{
a scatterplot
}
\description{
Graphs a scatterplot of gene-metabolite pairs vs. the interaction coefficient for the gene-metabolite pair
}
\examples{
\dontrun{
dir <- system.file("extdata", package="IntLIM", mustWork=TRUE)
csvfile <- file.path(dir, "NCItestinput.csv")
mydata <- ReadData(csvfile,metabid='id',geneid='id')
myres <- RunIntLim(mydata,stype="PBO_vs_Leukemia")
InteractionCoefficientGraph(inputResults=myres)
}
}
