% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{InteractionCoefficientGraph}
\alias{InteractionCoefficientGraph}
\title{Graphs a scatterplot of pairs vs. the interaction coefficient
for the pair}
\usage{
InteractionCoefficientGraph(
  inputResults,
  interactionCoeffPercentile = 0.1,
  percentageToPlot = 0.01,
  independent.var.type = 1,
  outcome = 2
)
}
\arguments{
\item{inputResults}{Data frame with model results (output of ProcessResults())}

\item{interactionCoeffPercentile}{percentile cutoff for interaction coefficient 
(default bottom 10 percent (high negative coefficients) and top 10 percent 
(high positive coefficients))}

\item{percentageToPlot}{percentage of points to plot (the points will be 
randomly selected) -- plotting all points will likely overwhelm plotting function.}

\item{independent.var.type}{type of analyte used as the independent variable 
("1" or "2")}

\item{outcome}{type of analyte used as the outcome/dependent variable ("1"
or "2")}
}
\value{
a scatterplot
}
\description{
Graphs a scatterplot of pairs vs. the interaction coefficient
for the pair
}
