% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{IntLimResults-class}
\alias{IntLimResults-class}
\title{IntLimResults class}
\description{
IntLimResults class
}
\section{Slots}{

\describe{
\item{\code{interaction.pvalues}}{matrix of interaction p-values}

\item{\code{interaction.adj.pvalues}}{matrix of adjusted interaction pvalues}

\item{\code{corr}}{matrix of correlations in group 1 and 2}

\item{\code{warnings}}{a message of whether genes and metabolites have 0 standard deviation}

\item{\code{stype}}{column name that represents sample type (by default, it will be used
in the interaction term). Only 2 categories are currently supported.}

\item{\code{outcome}}{outcome is either 'metabolite' or 'gene'}
}}

