% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{MarginalEffectsGraph}
\alias{MarginalEffectsGraph}
\title{Creates a dataframe of the marginal effect of phenotype}
\usage{
MarginalEffectsGraph(dataframe, title, ylab, xlab)
}
\arguments{
\item{dataframe}{from MarginalEffectsGraphDataframe}

\item{title}{for graph}

\item{ylab}{outcome analyte in pair}

\item{xlab}{independent analyte in pair}
}
\value{
values used for graphing
}
\description{
Creates a dataframe of the marginal effect of phenotype
}
