% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{MarginalEffectsGraphDataframe}
\alias{MarginalEffectsGraphDataframe}
\title{Creates a graph of the marginal effect of phenotype}
\usage{
MarginalEffectsGraphDataframe(inputResults, inputData, geneOfInterest,
  metaboliteOfInterest)
}
\arguments{
\item{inputResults}{IntLimResults object with model results (output of RunIntLim())}

\item{inputData}{MultiDataSet object (output of ReadData()) with gene expression,}

\item{geneOfInterest}{gene in gene-metabolite pair}

\item{metaboliteOfInterest}{metabolite in gene-metabolite pair}
}
\value{
dataframe for further analysis
}
\description{
Creates a graph of the marginal effect of phenotype
}
