% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{MarginalEffectsGraphDataframeMetabolitePairs}
\alias{MarginalEffectsGraphDataframeMetabolitePairs}
\title{Creates a dataframe of the marginal effect of phenotype}
\usage{
MarginalEffectsGraphDataframeMetabolitePairs(
  inputResults,
  inputData,
  metaboliteOfInterest1,
  metaboliteOfInterest2
)
}
\arguments{
\item{inputResults}{IntLimResults object with model results (output of RunIntLim())}

\item{inputData}{MultiDataSet object (output of ReadData()) with gene expression,}

\item{metaboliteOfInterest1}{outcome metabolite in metabolite pair}

\item{metaboliteOfInterest2}{independent metabolite in metabolite pair}
}
\value{
dataframe for further analysis
}
\description{
Creates a dataframe of the marginal effect of phenotype
}
