% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotDistributions}
\alias{PlotDistributions}
\title{Get some stats after reading in data}
\usage{
PlotDistributions(inputData, viewer = TRUE, palette = "Set1")
}
\arguments{
\item{inputData}{IntLimObject output of ReadData()}

\item{viewer}{whether the plot should be displayed in the RStudio viewer (TRUE) or
in Shiny/Knittr (FALSE)}

\item{palette}{choose an RColorBrewer palette ("Set1", "Set2", "Set3",
"Pastel1", "Pastel2", "Paired", etc.) or submit a vector of colors}
}
\value{
a highcharter object
}
\description{
Get some stats after reading in data
}
