% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotGMPair}
\alias{PlotGMPair}
\title{scatter plot of gene-metabolite pairs (based on user selection)}
\usage{
PlotGMPair(inputData, stype = NULL, geneName, metabName, palette = "Set1",
  viewer = T)
}
\arguments{
\item{inputData}{IntLimObject output of ReadData() or FilterData()}

\item{stype}{category to color-code by}

\item{geneName}{string of select geneName}

\item{metabName}{string of select metabName}

\item{palette}{choose an RColorBrewer palette ("Set1", "Set2", "Set3",
"Pastel1", "Pastel2", "Paired", etc.) or submit a vector of colors}

\item{viewer}{whether the plot should be displayed in the RStudio viewer (T) or
in Shiny/Knittr (F)}
}
\value{
a highcharter object
}
\description{
scatter plot of gene-metabolite pairs (based on user selection)
}
\examples{
\dontrun{
dir <- system.file("extdata", package="IntLim", mustWork=TRUE)
csvfile <- file.path(dir, "NCIinput.csv")
mydata <- ReadData(csvfile,metabid='id',geneid='id')
PlotGMPair(mydata,stype="PBO_vs_Leukemia","DLG4","(p-Hydroxyphenyl)lactic acid")

}
}
