% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotGraphWeightsHeatmap}
\alias{PlotGraphWeightsHeatmap}
\title{Plot the graph as a heatmap with edges colored by interaction coefficient.}
\usage{
PlotGraphWeightsHeatmap(inputResults, inputData)
}
\arguments{
\item{inputResults}{The IntLIM results (from RunIntLim())}

\item{inputData}{The input data (from ReadData())}
}
\description{
Plot the graph as a heatmap with edges colored by interaction coefficient.
}
