% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_ProcessResults.R
\name{ProcessResultsContinuous}
\alias{ProcessResultsContinuous}
\title{Retrieve significant gene-metabolite pairs (aka filter out nonsignificant pairs) based on value of gene:type interaction coefficient from linear model}
\usage{
ProcessResultsContinuous(inputResults, InteractionCoeffcutoff = 0.5,
  pvalcutoff = 0.05)
}
\arguments{
\item{inputResults}{IntLimResults object with model results (output of RunIntLim())}

\item{InteractionCoeffcutoff}{Smallest interaction coefficient that will be graphed (positive or negative)}

\item{pvalcutoff}{cutoff of FDR-adjusted p-value for filtering (default 0.05)}
}
\value{
IntLimResults object with model results (now includes filt.results data)
}
\description{
Retrieve significant gene-metabolite pairs (aka filter out nonsignificant pairs) based on value of gene:type interaction coefficient from linear model
}
