% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_ProcessResults.R
\name{ProcessResultsContinuous}
\alias{ProcessResultsContinuous}
\title{Retrieve significant gene-metabolite pairs (aka filter out nonsignificant pairs) based on value of gene:type interaction coefficient from linear model}
\usage{
ProcessResultsContinuous(inputResults, interactionCoeffPercentile = 0.1,
  pvalCutoff = 0.05)
}
\arguments{
\item{inputResults}{IntLimResults object with model results: output of RunIntLim}

\item{interactionCoeffPercentile}{percentile cutoff for interaction coefficient
default bottom 10 percent (high negative coefficients) and top 10 percent (high positive coefficients)}

\item{pvalCutoff}{cutoff of FDR-adjusted p-value for filtering (default 0.05)}
}
\description{
Retrieve significant gene-metabolite pairs (aka filter out nonsignificant pairs) based on value of gene:type interaction coefficient from linear model
}
