% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_ProcessResults.R
\name{ProcessResultsContinuous}
\alias{ProcessResultsContinuous}
\title{Retrieve significant pairs, based on adjusted p-values, interaction coefficient
percentile, and R^2 values for continuous models}
\usage{
ProcessResultsContinuous(
  inputResults,
  coeffPercentile = 0.1,
  pvalCutoff = 0.05,
  rsquaredCutoff = 0,
  coefficient = "interaction"
)
}
\arguments{
\item{inputResults}{IntLimResults object with model results (output of RunIntLim())}

\item{coeffPercentile}{percentile cutoff for absolute value of coefficient}

\item{pvalCutoff}{cutoff of FDR-adjusted p-value for filtering (default 0.05)}

\item{rsquaredCutoff}{cutoff for lowest r-squared value}

\item{coefficient}{Coefficient to filter by. Default is interaction coefficient.
Other options are "stype" (phenotype) and "analyte" (independent analyte).
Other options are "stype" and "analyte".}
}
\value{
A data frame with the following columns for each pair of analytes:
"Analyte1", "Analyte2", "interaction_coeff", "Pval", "FDRadjPval", and "rsquared".
Optionally, coefficients for each covariate may also be included.
}
\description{
Retrieve significant pairs, based on adjusted p-values, interaction coefficient
percentile, and R^2 values for continuous models
}
