% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_RunPairwisePrediction.R
\name{ProjectPredictionsOntoGraph}
\alias{ProjectPredictionsOntoGraph}
\title{Given a graph and a phenotype prediction for each significant pair, generate a 
new graph with the phenotype predictions included as the edge weights.}
\usage{
ProjectPredictionsOntoGraph(predictions, coRegulationGraph)
}
\arguments{
\item{predictions}{A matrix of predictions. Each signficant pair of analytes 
results in a prediction for each subject.}

\item{coRegulationGraph}{An igraph object. This graph is the co-regulation graph
generated using IntLIM analysis of analyte pairs.}
}
\description{
Given a graph and a phenotype prediction for each significant pair, generate a 
new graph with the phenotype predictions included as the edge weights.
}
