% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{pvalCorrVolcano}
\alias{pvalCorrVolcano}
\title{'volcano' plot (difference in correlations vs p-values) 
of all gene-metabolite pairs}
\usage{
pvalCorrVolcano(inputResults, inputData)
}
\arguments{
\item{inputResults}{IntLimResults object with model results (output of RunIntLim())}

\item{inputData}{MultiDataSet object (output of ReadData()) with gene expression,}
}
\value{
a smoothScatter plot
}
\description{
'volcano' plot (difference in correlations vs p-values) 
of all gene-metabolite pairs
}
\examples{
\dontrun{
csvfile <- file.path(dir, "NCItestinput.csv")
mydata <- ReadData(csvfile,metabid='id',geneid='id')
myres <- RunIntLim(mydata,stype="PBO_vs_Leukemia")
pvalCorrVolcano(myres,mydata)
}
}
