% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PValueBoxPlots}
\alias{PValueBoxPlots}
\title{Visualize the distribution of unadjusted p-values for all covariates
from linear models using a bar chart.}
\usage{
PValueBoxPlots(IntLimResults)
}
\arguments{
\item{IntLimResults}{output of RunIntLim()}
}
\value{
No return value, called for side effects
}
\description{
Visualize the distribution of unadjusted p-values for all covariates
from linear models using a bar chart.
}
