% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ReadData.R, R/03_RunIntLim.R
\name{ReadData}
\alias{ReadData}
\alias{ReadData}
\title{Read in CSV file}
\usage{
ReadData(inputFile, stype = NULL, outcome = "metabolite")

ReadData(inputFile, stype = NULL, outcome = "metabolite")
}
\arguments{
\item{inputFile}{input file in CSV format (see Despcription)}

\item{stype}{column name that represents sample type (by default, it will be used 
in the interaction term). Only 2 categories are currently supported.}

\item{outcome}{'metabolite' or 'gene' must be set as outcome/independent variable
(default is 'metabolite')}

\item{metabid}{name of column from metabolite meta data to be used as id
(required, must match metabolite abundances data)}

\item{geneid}{name of column from gene meta data to be used as id
(required, must match gene expression data)}

\item{logmetab}{whether or not to log metabolite values (T/F)}

\item{loggene}{whether or not to log gene values (T/F)}

\item{inputData}{MultiDataSet object (output of ReadData()) with gene expression,
metabolite abundances, and associated meta-data}
}
\value{
MultiDataSet object with input data

IntLimModel object with model results
}
\description{
The metadata associated with data files to be analyzed in IntLim is supplied
as a CSV file with two columns and 6 rows: 
   type,filenames
   metabData,myfilename
   geneData,myfilename
   metabMetaData,myfilename (optional)
   geneMetaData,myfilename (optional)
   sampleMetaData,myfilename

Run linear models and retrieve relevant statistics
}
\details{
Note that all files supplied in the CSV file, and the CSV file itself should be placed in the same folder.  The software assumes will automatically retrieve the file path of
the input files (based on location of CSV files).  
Note also that the input data files should be in a specific format:
metabData: rows are metabolites, columns are samples
geneData: rows are genes, columns are samples
metabMetaData: rows are metabolites, features are columns
geneMetaData: rows are genes, features are columns
sampleMetaData: rows are samples, features are columns
In addition, the first column of the sampleMetaData file is assumed to be the sample id, 
and those sample ids should match the columns of metabData and geneData (e.g. it is required
that all sample ids in the metabData and geneData are also in the sampleMetaDatafile).
}
\examples{
dir <- system.file("extdata", package="IntLim", mustWork=TRUE)
csvfile <- file.path(dir, "test.csv")
mydata <- ReadData(csvfile,metabid='BIOCHEMICAL',geneid='X')
dir <- system.file("extdata", package="IntLim", mustWork=TRUE)
csvfile <- file.path(dir, "test.csv")
mydata <- ReadData(csvfile,metabid='BIOCHEMICAL',geneid='X')
myres <- RunIntLim(mydata,stype="DIAG")
}
