% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ReadData.R
\name{ReadData}
\alias{ReadData}
\title{Read in CSV file}
\usage{
ReadData(
  inputFile,
  analyteType1id = "id",
  analyteType2id = "id",
  logAnalyteType1 = FALSE,
  logAnalyteType2 = FALSE,
  class.feat = list(),
  suppressWarnings = FALSE
)
}
\arguments{
\item{inputFile}{input file in CSV format (see Despcription)}

\item{analyteType1id}{name of column from Analyte Type 1 meta data to be used as id
(required if an Analyte Type 1 meta data file is present, 
must match Analyte Type 1 data)}

\item{analyteType2id}{name of column from Analyte Type 2 meta data to be used as id
(required if an Analyte Type 2 meta data file is present, 
must match Analyte Type 2 data)}

\item{logAnalyteType1}{whether or not to log values for Analyte Type 1(T/F)}

\item{logAnalyteType2}{whether or not to log values for Analyte Type 2(T/F)}

\item{class.feat}{class ("factor" or "numeric") for each covariate. The following format
is required: list(covar1="numeric", covar2="factor")}

\item{suppressWarnings}{whether or not to suppress warnings}
}
\value{
IntLimData object with input data
}
\description{
Read in CSV file
}
