% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_RunIntLim.R
\name{RunIntLim}
\alias{RunIntLim}
\title{Run linear models and retrieve relevant statistics}
\usage{
RunIntLim(inputData, stype = NULL, outcome = "metabolite")
}
\arguments{
\item{inputData}{MultiDataSet object (output of ReadData()) with gene expression,
metabolite abundances, and associated meta-data}

\item{stype}{column name that represents sample type (by default, it will be used 
in the interaction term). Only 2 categories are currently supported.}

\item{outcome}{'metabolite' or 'gene' must be set as outcome/independent variable
(default is 'metabolite')}
}
\value{
IntLimModel object with model results
}
\description{
Run linear models and retrieve relevant statistics
}
\examples{
dir <- system.file("extdata", package="IntLim", mustWork=TRUE)
csvfile <- file.path(dir, "test.csv")
mydata <- ReadData(csvfile,metabid='id',geneid='id')
myres <- RunIntLim(mydata,stype="cancertype")
}
