% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_RunIntLim.R
\name{RunIntLim}
\alias{RunIntLim}
\title{Run linear models and retrieve relevant statistics}
\usage{
RunIntLim(
  inputData,
  stype = NULL,
  outcome = "metabolite",
  covar = NULL,
  continuous = FALSE,
  save.covar.pvals = FALSE,
  independent.var.type = "gene",
  remove.duplicates = FALSE
)
}
\arguments{
\item{inputData}{Named list (output of 
FilterData()) with gene expression, metabolite abundances, 
and associated meta-data}

\item{stype}{column name that represents sample type (by default, it will be used
in the interaction term). Only 2 categories are currently supported.}

\item{outcome}{'metabolite' or 'gene' must be set as outcome/independent variable
(default is 'metabolite')}

\item{covar}{Additional variables from the phenotypic data that be integrated into linear model}

\item{continuous}{boolean to indicate whether the data is continuous or discrete}

\item{save.covar.pvals}{boolean to indicate whether or not to save the p-values of all covariates,
which can be analyzed later but will also lengthen computation time. The default is FALSE.}

\item{independent.var.type}{'metabolite' or 'gene' must be set as independent variable
(default is 'metabolite')}

\item{remove.duplicates}{boolean to indicate whether or not to remove the metabolite-metabolite
or gene-gene pair with the highest p-value across two duplicate models (e.g. m1~m2 and m2~m1)}
}
\value{
IntLimModel object with model results
}
\description{
Run linear models and retrieve relevant statistics
}
