% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_LearnGraphPredictionModel.R
\name{RunLeastSquaresOptimization}
\alias{RunLeastSquaresOptimization}
\title{Run least squares optimization to optimize the weights}
\usage{
RunLeastSquaresOptimization(modelInput, convolution)
}
\arguments{
\item{modelInput}{An object of the ModelInput class.}

\item{convolution}{Whether or not to convolve the input.}
}
\description{
Run least squares optimization to optimize the weights
}
