% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{RunLM}
\alias{RunLM}
\title{Function that runs linear models and returns interaction p-values.}
\usage{
RunLM(incommon, outcome = "metabolite", type = NULL, covar = NULL)
}
\arguments{
\item{incommon}{MultiDataSet object (output of ReadData()) with gene}

\item{outcome}{'metabolite' or 'gene' must be set as outcome/independent variable 
(default is 'metabolite')}

\item{type}{vector of sample type (by default, it will be used in the interaction term).
Only 2 categories are currently supported.}

\item{covar}{vector of additional vectors to consider}
}
\description{
Function that runs linear models and returns interaction p-values.
}
