% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsfunctions.R
\name{ShowStats}
\alias{ShowStats}
\title{Get some stats after reading in data}
\usage{
ShowStats(IntLimObject)
}
\arguments{
\item{IntLimObject}{output of ReadData()}
}
\value{
data.frame with some # of samples, features, etc.
}
\description{
Get some stats after reading in data
}
