% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_LearnGraphPredictionModel.R
\name{TrainGraphLearningModel}
\alias{TrainGraphLearningModel}
\title{Train the graph learning model, using the specifications in the ModelResults.
class and storing the results in the ModelResults class.}
\usage{
TrainGraphLearningModel(modelResults)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}
}
\description{
Train the graph learning model, using the specifications in the ModelResults.
class and storing the results in the ModelResults class.
}
